% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regts.R
\name{adjust_period}
\alias{adjust_period}
\title{Adjust the period of the timeseries}
\usage{
adjust_period(x, range)
}
\arguments{
\item{x}{a timeseries object (\code{ts} or \code{regts})}

\item{range}{the new period range of the timeseries}
}
\value{
a \code{regts} with the adjusted period
}
\description{
Adjust the period of the timeseries
}

