% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod.R
\name{as.character.regperiod}
\alias{as.character.regperiod}
\title{Convert an \code{regperiod} to a character representation}
\usage{
\method{as.character}{regperiod}(x)
}
\description{
param x the \code{regperiod} object to be converted
}

