% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{as.list.regts}
\alias{as.list.regts}
\title{Convert a regts to a list of univariate regts objects}
\usage{
\method{as.list}{regts}(x, ...)
}
\arguments{
\item{x}{a \code{regts} object}

\item{...}{arguments passed to methods (not used in the default implementation)}
}
\value{
a list of univariate \code{regts} objects
}
\description{
Convert a regts to a list of univariate regts objects
}
\examples{
ts <- regts(matrix(1:6, ncol = 2), start = "2015Q3", names = c("a", "b"))
print(as.list(ts))
}

