% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{as.period}
\alias{as.period}
\alias{as.period.character}
\alias{as.period.numeric}
\title{Coerce an R object to a \code{period} object if possible.}
\usage{
as.period(x, ...)

\method{as.period}{character}(x, frequency = NA, ...)

\method{as.period}{numeric}(x, frequency = NA, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{additional arguments to be passed to or from methods (cuyrrently
not used in package \code{regts})}

\item{frequency}{frequency of the period object.
Argument \code{frequency} is mandatory if a general period format
such as \code{"2011-1"} has been specified}
}
\value{
a \code{period} object
}
\description{
This function can coerce a character string or numeric to a
\code{period}. For details see the documentation of function \code{\link{period}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Convert a character string to a \code{period}

\item \code{numeric}: Convert a numeric to a \code{period}
}}

\examples{
as.period("2010Q3")
as.period(2010)
}
\seealso{
\code{\link{period}} and \code{\link{as.period_range}}
}
