% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_range.R
\name{as.period_range}
\alias{as.period_range}
\alias{as.period_range.character}
\alias{as.period_range.period}
\title{Coerce an R object to a \code{period_range} object if possible.}
\usage{
as.period_range(x, frequency = NA, ...)

\method{as.period_range}{character}(x, frequency = NA, ...)

\method{as.period_range}{period}(x, ...)
}
\arguments{
\item{x}{an R object}

\item{frequency}{frequency of the period object.
Argument \code{frequency} is mandatory if a general period format
such as \code{"2011-1"} has been specified}

\item{...}{additional arguments to be passed to or from methods (cuyrrently
not used in package \code{regts})}
}
\value{
a \code{period_range} object
}
\description{
This function can coerce a character string to a
\code{period_range}. For details see the documentation of function
\code{\link{period_range}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Convert a character string to a
\code{period_range}

\item \code{period}: Convert a \code{period} to a \code{period_range}
}}

\examples{
# create a period_range for a timeseries with frequency 2 (half year)
period_range("2010-2", "2016-2", frequency = 2)

# convert a period object to a period_range:
p <- period("2010Q2")
as.period_range(p)
}
\seealso{
\code{\link{period_range}} and \code{\link{period}}
}
