% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod_range.R
\name{as.regperiod_range.character}
\alias{as.regperiod_range.character}
\title{Convert a character string to a regperiod_range object}
\usage{
\method{as.regperiod_range}{character}(x, frequency = NA, ...)
}
\arguments{
\item{x}{a character string}

\item{frequency}{(mandatory if a period format without frequency indicator
has been used, e.g. \code{"2011-3"})}
}
\description{
Convert a character string to a regperiod_range object. The first and last
period of the range are specied as in \link{regts}. The two periods
are separated with \code{"/"} (e.g. \code{"2010Q2/2016Q2"}).
The first or last period may be omitted: in that case the period range
has no lower of upper bound (e.g. \code{"2012Q3/"}).
}
\examples{
as.regperiod_range("2010Q2/2016Q3")
as.regperiod_range("2010Q2/")
as.regperiod_range("/2016Q3")

# a single period can also be converted to a regperiod_range
as.regperiod_range("2016Q1")
}

