% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod_range.R
\name{as.regperiod_range}
\alias{as.regperiod_range}
\alias{as.regperiod_range.character}
\alias{as.regperiod_range.regperiod}
\title{Coerce an R object to a \code{\link{regperiod_range}}}
\usage{
as.regperiod_range(x, frequency = NA, ...)

\method{as.regperiod_range}{character}(x, frequency = NA, ...)

\method{as.regperiod_range}{regperiod}(x, ...)
}
\arguments{
\item{x}{an R object}

\item{frequency}{the frequency (mandatory for
\code{as.regperiod_range.character}
if a period format without frequency indicator has been used,
e.g. \code{"2011-3"})}

\item{...}{arguments passed to methods (not used in this package)}
}
\value{
a \code{regperiod_range}
}
\description{
Coerce an R object to a \code{\link{regperiod_range}}
}
\details{
\code{as.regperiod_range.character} converts a character string to a
regperiod_range object. The format for the first and last period of
the range are specied as in \code{\link{regperiod}}. The two periods strings
are separated by \code{"/"} (e.g. \code{"2010Q2/2016Q2"}).
The first or last period may be omitted: in that case the period range
has no lower of upper bound (e.g. \code{"2012Q3/"}). The \code{"/"}
separator is also optional: if omitted, then the period range
contains a single period.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Convert a character string to a
regperiod_range object

\item \code{regperiod}: Convert a single \code{\link{regperiod}} to a
\link{regperiod_range}.
}}
\examples{
as.regperiod_range("2010Q2/2016Q3")
as.regperiod_range("2010Q2/")
as.regperiod_range("/2016Q3")
as.regperiod_range("2016Q1")

# convert a regperiod object to a regperiod_range:
p <- regperiod("2010Q2")
as.regperiod_range(p)
}

