% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regts.R
\name{as.regts}
\alias{as.regts}
\title{Coerce an object to a \link{regts} timeseries object}
\usage{
as.regts(x)
}
\arguments{
\item{x}{an arbitrary R object}
}
\value{
a \link{regts} object
}
\description{
Coerce an object to a \link{regts} timeseries object
}

