% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{as.ts.regts}
\alias{as.ts.regts}
\title{Convert a \link{regts} to a normal \link{ts}}
\usage{
\method{as.ts}{regts}(x, ...)
}
\arguments{
\item{x}{a \code{regts}}
}
\value{
a \code{ts}
}
\description{
Convert a \link{regts} to a normal \link{ts}
}

