% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_matrix.R
\name{as_matrix}
\alias{as_matrix}
\alias{as_matrix.ts}
\title{Convert a \code{ts} to a \code{matrix}}
\usage{
as_matrix(x, ...)

\method{as_matrix}{ts}(x, rowwise = FALSE, ...)
}
\arguments{
\item{x}{a \code{\link[stats]{ts}} or \code{\link{regts}}}

\item{...}{additional arguments to be passed to methods.}

\item{rowwise}{a logical value: should the timeseries be stored rowwise
or columnwise in the matrix? Defaults to \code{FALSE}}
}
\value{
A \code{\link[base]{matrix}}
}
\description{
This function converts a \code{\link[stats]{ts}} object
to a normal  \code{\link[base]{matrix}}, i.e. a matrix without timeseries
class and attributes.The periods are stored in the row or
column names of the returned matrix, depending on argument \code{rowwise}.
}
\details{
The function behaves differently than base method
\code{\link[base]{as.matrix}}. If the input timeseries is a matrix with
timeseries attributes, then \code{as.matrix} just returns the input value.
If the input timeseries is not a matrix (a univariate timeseries with vector
data), then \code{as.matrix} returns a matrix without row and column names
and without timeseries attributes. In contrast, \code{as_matrix} always
returns a matrix without timeseries attributes.
}
\section{Methods (by class)}{
\itemize{
\item \code{ts}: Coerce a \code{\link[stats]{ts}} to a matrix without
timeseries class and attributes
}}

\examples{
ts <- regts(matrix(1:4, ncol = 2) , start = 2015, names = c("a", "b"))
as_matrix(ts, rowwise = TRUE)
}
