% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_frequency.R
\name{change_frequency}
\alias{change_frequency}
\alias{change_frequency.period}
\alias{change_frequency.period_range}
\title{Change the frequency of a \code{\link{period}} or \code{\link{period_range}}
object.}
\usage{
change_frequency(x, new_frequency, ...)

\method{change_frequency}{period}(x, new_frequency, ...)

\method{change_frequency}{period_range}(x, new_frequency, ...)
}
\arguments{
\item{x}{a \code{\link{period}} or \code{\link{period_range}}}

\item{new_frequency}{the new_frequency}

\item{...}{arguments passed to methods (not used in package \code{regts})}
}
\value{
a \code{period} or \code{period_range} (depending on the type
of argument \code{x}) with the new frequency
}
\description{
A \code{period} can be converted to lower frequency. For example,
a month \code{"2017M4"} can be converted to the quarter \code{"2017Q2"}.
The old frequency should be divisible by the new frequency.
\cr\cr
A \code{period_range} can be converted to both lower and higher frequency.
For example, the range \code{"2017Q3/2018Q1"} can be converted to
the month range \code{"2017M7/2018M3"} or the year range
\code{"2017/2018"}.
If the \code{period_range} is converted to lower frequency, then the
old frequency should be divisible by the new frequency.
If the range is converted to higher frequency, then
the new frequency should be divisible by the old frequency.
}
\section{Methods (by class)}{
\itemize{
\item \code{change_frequency(period)}: Change the frequency of a \code{period} to
lower frequency

\item \code{change_frequency(period_range)}: Change the frequency of a \code{period_range} to
higher or lower frequency

}}
\examples{
p <- period("2017M4")
change_frequency(p, 4)

range <- period_range("2017Q3/2018Q1")
change_frequency(range, 12)
change_frequency(range, 1)
}
