% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disagg.R
\name{disagg}
\alias{disagg}
\title{Temporal disaggregation of time series using spline interpolation.}
\usage{
disagg(x, nfrequency, method = c("fmm", "periodic", "natural", "monoH.FC",
  "hyman"), constraint = c("sum", "average", "last", "first"))
}
\arguments{
\item{x}{a \code{\link[stats]{ts}} of \code{\link{regts}} object}

\item{nfrequency}{the frequency of the result. This should be higher than
the frequency of timeseries \code{x}.}

\item{method}{specifies the type of spline to be used.
Possible values are \code{"fmm"}, \code{"natural"}, \code{"periodic"},
\code{"monoH.FC"} and \code{"hyman"}. Can be abbreviated.}

\item{constraint}{Constraint on the high frequency result.
Possible values are \code{"sum"}, \code{"average"}, \code{"last"},
\code{"first"}.
Either the sum, the average, the first or the last value of the resulting
high-frequency series should be equal to the corresponding
low frequency value}
}
\description{
This function converts an timeseries to a timeseries with higher frequency,
for example a yearly timeseries to a quartely timeseries. Missing values
are obtained by spline interpolation. This function employs function
\code{\link[stats]{spline}} of the \code{stats} package
}
\details{
For each individual timeseries, trailing and leading \code{NA} values are
removed before interpolation. If the timeseries contains intermediate
\code{NA} values, the the resulting timeseries is set to NA.

For details about the different spline methods,
consult the documentation of the \code{\link[stats]{spline}}
function of the \code{stats} package.
}
