% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_end_period.R
\name{end_period}
\alias{end_period}
\title{Returns the end period of a \code{\link{regts}} or a
\code{\link{regperiod_range}}}
\usage{
end_period(x)
}
\arguments{
\item{x}{a \code{regts} or \code{regperiod_range} object}
}
\value{
a \code{regperiod} object representing the last period of the
range. The return value can be \code{NULL} if argument \code{x} is a
\code{regperiod_range} with no upper boundary.
}
\description{
Returns the end period of a \code{\link{regts}} or a
\code{\link{regperiod_range}}
}

