% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_replace.R
\name{extract}
\alias{extract}
\alias{[}
\alias{[[}
\alias{[<-}
\alias{$}
\alias{$<-}
\title{Extract of Replace Parts of a \code{regts} object.}
\usage{
# Select parts of a regts.
x[i]
x[i, j, drop = TRUE]
x$name

# Select parts of a regts as a plain vector or matrix:
x[[i]]
x[[i, j, drop = TRUE]]

x[i] <- value
x[i, j] <- value

x$name <- value
}
\arguments{
\item{x}{A \code{regts} timeseries object.}

\item{i}{Indices for the first dimension.
A numeric vector or a \code{\link{period}},
\code{\link{period_range}} or an character that be coerced to
a \code{period} or \code{period_range} object. The frequency of the
specified period or period range must be equal to or lower than
the frequency of the timeseries.}

\item{j}{A numeric vector with column indices or a character vector with
column names. Only used in the underlying data of the timeseries
is a matrix.}

\item{drop}{Relevant for extraction if the underlying data of the
timeseries is a  matrix and if argument \code{j} has length 1. In
that case, the timeseries is converted to a vector timeseries
(dropping the column names) if \code{drop = TRUE}.
This argument is only used for extraction, not for
replacement. Specify \code{drop = FALSE} if you want to retain
the column names.}

\item{value}{The replacement value(s), usually a vector, matrix or
another timeseries object.}

\item{name}{A rel}
}
\description{
Operators acting on regts objects to extract or replace parts.
}
\section{Details}{


The \code{[} operator can be used to extact par:qt of a \code{regts} object. When
combined with the assignment operator \verb{<-}, it can also be used to
replace part of a regts object, or to extend a regts (see the examples
below).
For \code{regts} objects, the \code{[[} works similarly as the \code{[}
operator, except that it always returns a normal vector or matrix and not a
timeseries.

\strong{Vector timeseries}

A vector timeseries is  a univariate timeseries for which the underlying
data is a vector (and not a matrix). A vector timeseries only has one
dimension and only argument \code{i} is relevant.

If \code{i} is a numeric vector, the result is a vector (not a timeseries)
with the observations at the specified indices.
If \code{i} is a \code{period}, \code{period_range}, or a character that can
be coerced to a \code{period} or \code{period_range},  the result is a timeseries
with a period equal to the specified period (range). The result is padded
with NA values if the specified period or period range lies outside the
period range of the input timeseries.

The result of the \code{[[} operator is always a vector.

\strong{Matrix timeseries}

For a matrix timeseries the underlying data is a matrix, usually with
column  names. The first selection argument \code{i} can be used to
select rows or a period range, similarly as argument \code{i} is used the
select elements of vector timeseries.
If \code{i} is a numerical vector, the result is a normal matrix (not a
timeseries), if \code{i} is a \code{period} or \code{period_range} the result is a
timeseries.

For matrix timeseries, argument \code{j} can be used to select columns.
When used in combination with the \verb{<-} operator I can be used to add
columns. If \code{j} has length 1, the result is a vector timeseries
unless \code{drop = FALSE}.
The result of the \code{[[} operator is always a plain matrix (not a timeseries).
}

\examples{

# univariate timeseries
#######################

x <- regts(1:5, start = "2018q1")
x[c(1, 3)]
x["2018q2"] # select period 2018q2
x[["2018q2"]] # get value in 2018q2 as numeric,

x["2018"]  # select all quarters in the year 2018
x[["2018"]] # get values for the year 2018 as numeric vector

# extend the period of the timeseries:
x["2020q1"] <- 12
x

# multivariate timeseries
#######################

x <- regts(matrix(1:10, ncol = 2), names = c("a", "b"), start = "2018q1")

x[c(1, 3)]
x["2018q2"] # select period 2018q2
x[["2018q2"]] # get value in 2018q2 as numeric,"

x["2018", "b"]   # select all quarters of column b in the year 2018
x[["2018", "b"]] # get values for the year 2018 as numeric vector

# now keep column names
x["2018", "b", drop = FALSE]
x[["2018", "b", drop = FALSE]]

# Extend the period range of the timeseries:
x["2019q3"] <- 12
x

# Extend the period range and add a new column
x["2019q4", "c"] <- 12
x

x$b   # Select columb b

x$d <- 2  # Add a new column d, with value 2
x
}
