% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_columns}
\alias{filter_columns}
\title{Select columns using a filter}
\usage{
filter_columns(x, regex, drop = TRUE)
}
\arguments{
\item{x}{any R object with column names (\code{data.frame},
\code{matrix}, \code{ts}, \code{regts})}

\item{regex}{a regular expression used to select column}

\item{drop}{if \code{TRUE}, the result is coerced to a vector if possible.}
}
\value{
the selection of object \code{x}
}
\description{
This function selects columns of an R object that has columns names
(e.g. a \code{data.frame}, \code{matrix}, \code{ts} or
\code{regts})  using a
regular expression.
}

