% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{frequency.period}
\alias{frequency.period}
\title{Returns the frequency of a \link{period} object}
\usage{
\method{frequency}{period}(x, ...)
}
\arguments{
\item{x}{a \code{period}}

\item{...}{additional arguments for future methods}
}
\value{
the frequency of the \code{period}
}
\description{
Returns the frequency of a \link{period} object
}
\examples{
p <- period("2016Q1")
freq <- frequency(p)
}
