% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_range.R
\name{frequency.period_range}
\alias{frequency.period_range}
\title{Returns the frequency of a \link{period_range} object}
\usage{
\method{frequency}{period_range}(x, ...)
}
\arguments{
\item{x}{a \code{\link{period_range}}}

\item{...}{additional arguments for future methods}
}
\value{
the frequency of the \code{\link{period_range}}
}
\description{
Returns the frequency of a \link{period_range} object
}
\examples{
p <- period_range("2016Q1", "2018Q2")
freq <- frequency(p)
}
