% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{frequency}
\alias{frequency}
\alias{frequency.period}
\alias{frequency.period_range}
\title{Return the frequency of a \code{\link{period}} or \code{\link{period_range}}}
\usage{
\method{frequency}{period}(x, ...)

\method{frequency}{period_range}(x, ...)
}
\arguments{
\item{x}{a \code{\link{period}} or a \code{\link{period_range}}}

\item{...}{additional arguments for future methods}
}
\value{
the frequency of the \code{period} or the
\code{period_range}
}
\description{
Function returns the frequency of a \code{period} or a
\code{period_range}
}
\examples{
p <- period("2016Q1")
freq <- frequency(p)

p <- period_range("2016Q1", "2018Q2")
freq <- frequency(p)#'

}
