% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod.R
\name{frequency.regperiod}
\alias{frequency.regperiod}
\title{Returns the frequency of a \link{regperiod} object}
\usage{
\method{frequency}{regperiod}(x)
}
\arguments{
\item{x}{a \code{regperiod}}
}
\value{
the frequency of the period
}
\description{
Returns the frequency of a \link{regperiod} object
}

