% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod_range.R
\name{get_end_period}
\alias{get_end_period}
\title{Get the end period of the \link{regperiod_range}}
\usage{
get_end_period(x)
}
\arguments{
\item{x}{a \link{regperiod_range{ object}}}
}
\value{
a \link{regperiod} object representing the last period of the
range, or \code{NULL} if the range has no upper boundary.
}
\description{
Get the end period of the \link{regperiod_range}
}

