% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_period_range}
\alias{get_period_range}
\title{Return the \link{period_range} of a timeseries.}
\usage{
get_period_range(x)
}
\arguments{
\item{x}{a timeseries (\link{ts} or \link{regts})}
}
\value{
a \code{period_range}
}
\description{
Return the \link{period_range} of a timeseries.
}
\seealso{
\code{\link{get_periods}}
}
