% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regts.R
\name{get_regperiod_range}
\alias{get_regperiod_range}
\title{Returns the period range of the time series as a \code{regperiod_range} object}
\usage{
get_regperiod_range(x)
}
\arguments{
\item{x}{a \code{regts}}
}
\value{
a \code{regperiod_range}
}
\description{
Returns the period range of the time series as a \code{regperiod_range} object
}

