% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod_range.R
\name{get_start_period}
\alias{get_start_period}
\title{Get the start period of the \link{regperiod_range}}
\usage{
get_start_period(x)
}
\arguments{
\item{x}{a \link{regperiod_range{ object}}}
}
\value{
a \link{regperiod} object representing the first period of the
range, or \code{NULL} if the range has no lower boundary.
}
\description{
Get the start period of the \link{regperiod_range}
}

