% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{get_year}
\alias{get_year}
\title{Return the year of a \code{\link{period}}}
\usage{
get_year(x)
}
\arguments{
\item{x}{a \code{\link{period}}}
}
\value{
the year
}
\description{
Return the year of a \code{\link{period}}
}
\examples{
get_year(period("2010Q3"))
}
\seealso{
\code{\link{get_subperiod}}
}
