% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_ts.R
\name{growth_ts}
\alias{growth_ts}
\title{Return the growth function of a timeseries}
\usage{
growth_ts(x, lag = 1, keep_range = TRUE)
}
\arguments{
\item{x}{a \code{\link[stats]{ts}} or \code{\link{regts}} object}

\item{lag}{an integer indicating which lag to use}

\item{keep_range}{If \code{TRUE} (the default), then  the output
timeseries has the same period range as the input timeseries.
Then the result timeseries will have \code{lag} NA values at the start.
If \code{FALSE} then the result timeseries is \code{lag} periods
shorter than the input timeseries.}
}
\value{
a \code{regts} object with growth values
}
\description{
Function \code{growth_ts} computes the growth function of a timeseries.
This function is defined as: (x[t] - x[t-lag]) / |x[t-lag]|
}
\examples{
x <- regts(rnorm(10), start = "2018Q1")

growth_ts(x, lag = 4)
}
