% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_tail.R
\name{head.regts}
\alias{head.regts}
\title{Printing head of a regts}
\usage{
\method{head}{regts}(x, method = c("first", "last"), size = 10, ...)
}
\arguments{
\item{method}{if \code{first} then the first part of the timeseries is printed,
otherwise (\code{last}) the last part of the timeseries is printed.}

\item{size}{is the number of timeseries that are printed. Default = 10.}

\item{...}{arguments passed to \code{\link{head}}.}
}
\description{
With this function the head of a regts can be printed. The generic function
\code{\link{head}} is extended with parameters \code{method} and \code{size}.
}
\examples{
data <- regts(matrix(1:200, ncol = 20), start = "2010Q2",
              names = paste0("abc", 1:20))
head(data)
head(data, method = "last", size = 5)
}
\seealso{
\code{\link{tail.regts}}
}
