% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_ts.R
\name{index_ts}
\alias{index_ts}
\title{Construct an index timeseries by scaling}
\usage{
index_ts(x, base_period = start_period(get_period_range(x)),
  index_value = 100)
}
\arguments{
\item{x}{a \code{\link[stats]{ts}} of \code{\link{regts}} object}

\item{base_period}{The base period: a \code{\link{period}} or a
\code{\link{period_range}}, or an object that can be coerced to a period or
period range. By default base period is the first period of the input
timeseries.}

\item{index_value}{the index value (by default 100). If \code{base_period}\
is a single period, then the returned timeseries will be equal to the index value
at this period. If \code{base_period} is a period
range with multiple periods, then the average of the index timeseries
is equal to the index value}
}
\value{
an index timeseries
}
\description{
This function scales a timeseries by dividing all observations with
one selected observation or by the mean of a range of observations.
}
