% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod.R
\name{is.period}
\alias{is.period}
\title{Test if an object is a period.}
\usage{
is.period(x)
}
\arguments{
\item{x}{any R object}
}
\value{
\code{TRUE} if the object is a \code{\link{period}}
}
\description{
Test if an object is a period.
}
\examples{
p <- period("2016Q1")
is.period(p)
is.period("2016Q1")
}
