% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod_range.R
\name{is.period_range}
\alias{is.period_range}
\title{Test if an object is a period_range.}
\usage{
is.period_range(x)
}
\arguments{
\item{x}{any R object}
}
\value{
\code{TRUE} if the object is a \code{\link{period_range}}
}
\description{
Test if an object is a period_range.
}
\examples{
range <- period_range("2016Q1/2017Q1")
is.period_range(range)
is.period_range("2016Q1/2017Q1")
}
