% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod.R
\name{is.regperiod}
\alias{is.regperiod}
\title{Test if an object is a regperiod.}
\usage{
is.regperiod(x)
}
\arguments{
\item{x}{an object}
}
\value{
<code>TRUE</code> if the object a \code{regperiod}
}
\description{
Test if an object is a regperiod.
}

