% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regts.R
\name{is.regts}
\alias{is.regts}
\title{Tests whether an object is a \link{regts} timeseries object}
\usage{
is.regts(x)
}
\arguments{
\item{x}{an arbitrary R object}
}
\value{
\code{TRUE} if \code{x} is a \link{regts}
}
\description{
Tests whether an object is a \link{regts} timeseries object
}
\seealso{
\link{regts} \link{as.regts}
}

