% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_ts.R
\name{join_ts}
\alias{join_ts}
\title{Join a timeseries object with another timeseries object}
\usage{
join_ts(x1, x2, method = c("mult", "add"))
}
\arguments{
\item{x1}{the first timeseries (a \code{\link{regts}} or
\code{\link[stats]{ts}} object).}

\item{x2}{the second timeseries (a \code{regts} or \code{ts} object).}

\item{method}{two different ways to join the timeseries.
By default the timeseries are joined multiplicatively. This behaviour can be
changed by using method \code{add}.}
}
\value{
a \code{\link{regts}} object.
}
\description{
This function can be used to join a \code{(reg)ts}
object with another \code{(reg)ts} object.
The result is a joined \code{\link{regts}} object.
}
\details{
This function creates a new timeseries from two (partially) overlapping
timeseries. The two timeseries must have the same frequency.
The first timeseries must contain the most recent data.
All observations from the second timeseries are scaled in such a way that
the overlapping observations from the two timeseries have the same value
(on average). Scaling methods are:

\code{mult} multiplicative joining (default)

\code{add} additive joining

The period range of the result is the union of the period ranges of the
first and second timeseries.

In case of multivariate regts only the common columns are joined.
The non overlapping columns in both timeseries are added to the result.
If both input timeseries are vectors (no column names), the result is also a vector.
}
\examples{
x1 <- regts(matrix(data = rep(1:9), nc = 3), period = "2000/2002",
            names = c("a", "b", "c"))
x2 <- regts(matrix(data = rep(10:15), nc = 3), period = "2000/2001",
            names = c("a", "c", "d"))
join_ts(x1, x2, method = "add", period = "first")

}
\seealso{
\code{\link{regts}}
}
