% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag2_diff2.R
\name{lag_ts-lead_ts}
\alias{lag_ts-lead_ts}
\alias{lag_ts}
\alias{lead_ts}
\title{Lag or Lead a Timeseries}
\usage{
lag_ts(x, k = 1, keep_range = TRUE)

lead_ts(x, k = 1, keep_range = TRUE)
}
\arguments{
\item{x}{A univariate or multivariate timeseries}

\item{k}{The number of lags or leads (in units of observations).
Must be a positive number.}

\item{keep_range}{If \code{TRUE} (the default), then  the output
timeseries has the same period range as the input timeseries.
Then the result timeseries will have \code{k} \code{NA} values at the
beginning (\code{lead_ts}) or the end (\code{lag_ts}). For
If \code{FALSE} then the result timeseries has a shifter period range.}
}
\description{
Computes the lag or lead of a timeseries by shifting the observations
by a given number of periods.
}
\section{Functions}{
\itemize{
\item \code{lag_ts}: Lag a timeseries

\item \code{lead_ts}: Lead a timeseries
}}

\examples{
x <- regts(1:10, start = "2018q3")
lag_ts(x, 1)    # calculate x[t+1]
lead_ts(x, 1)   # return x[t-1]
}
