% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_ts_diff_ts.R
\name{lag_ts/lead_ts}
\alias{lag_ts/lead_ts}
\alias{lag_ts}
\alias{lead_ts}
\title{Lag or Lead a Timeseries}
\usage{
lag_ts(x, n = 1, keep_range = TRUE)

lead_ts(x, n = 1, keep_range = TRUE)
}
\arguments{
\item{x}{a univariate or multivariate timeseries}

\item{n}{the number of lags or leads (in units of observations).
Must be a positive number.}

\item{keep_range}{if \code{TRUE} (the default), then the output
timeseries has the same period range as the input timeseries.
Then the result timeseries will have \code{n} \code{NA} values at the
beginning (\code{lag_ts}) or the end (\code{lead_ts}).
If \code{FALSE} the period range of the result timeseries is shifted
by \code{n} periods. For example, for \code{lag_ts} the result timeseries
starts and ends \code{n} periods later.}
}
\description{
Compute the lag or lead of a timeseries, shifting the observations
by a given number of periods.
}
\details{
\code{lag_ts} and \code{lead_ts} differ from \code{\link[stats]{lag}}
in the \code{stats} package in that the specified number of lags or leads
is positive, and that by default the resulting timeseries has the same
period range as the input timeseries.
}
\section{Functions}{
\itemize{
\item \code{lag_ts}: Lag a timeseries

\item \code{lead_ts}: Lead a timeseries
}}

\examples{
x <- regts(1:10, start = "2018q3")
lag_ts(x, 1)                        # calculate x[t-1]
lead_ts(x, 1, keep_range = FALSE)   # calculate x[t+1]
}
