% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_ts_diff_ts.R
\name{lag_ts}
\alias{lag_ts}
\alias{lag_ts.ts}
\title{Lag a Timeseries}
\usage{
lag_ts(x, n = 1, keep_range = TRUE, ...)

\method{lag_ts}{ts}(x, n = 1, keep_range = TRUE, ...)
}
\arguments{
\item{x}{a univariate or multivariate timeseries.
Can also be a vector, matrix or data frame (see details).}

\item{n}{the number of lags (in units of observations).
Must be a positive number.}

\item{keep_range}{if \code{TRUE} (the default), then the output
timeseries has the same period range as the input timeseries.
The result timeseries will have \code{n} \code{NA} values at the
beginning. If \code{FALSE} the period range of the result timeseries is
shifted by \code{n} periods. The result timeseries
starts and ends \code{n} periods later.}

\item{...}{further arguments to be passed to or from methods
(currently not used in package \code{regts})}
}
\description{
Compute the lag of a timeseries, shifting the observations
forwards by a given number of periods.
}
\details{
Vector, matrix and data frame arguments are first converted to a \code{regts}
with function \code{\link{regts}}. This conversion results in a timeseries
with frequency 1 and starting at year 1.

\code{lag_ts} differs from \code{\link[stats]{lag}}
in the \code{stats} package in that the specified number of lags or leads
is positive, and that by default the resulting timeseries has the same
period range as the input timeseries.
}
\examples{
x <- regts(1:10, start = "2018q3")
lag_ts(x)
lag_ts(x, k = 2, keep_range = FALSE)
}
\seealso{
\code{\link{lead_ts}} and \code{\link{diff_ts}}
}
