% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod_range.R
\name{length_range}
\alias{length_range}
\title{Returns the number of periods in a \code{\link{regperiod_range}} object.}
\usage{
length_range(x)
}
\arguments{
\item{x}{a \code{regperiod_range}}
}
\value{
The number of periods in the range, or \code{Inf} is the
range is not bounded
}
\description{
Returns the number of periods in a \code{\link{regperiod_range}} object.
}
\examples{
range <- regperiod_range("2010Q2", "2011Q3")
length_range(range)  # the result will be 6
}
