% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod_range.R
\name{lensub}
\alias{lensub}
\title{Returns the number of subperiods in a \link{regperiod_range} object.}
\usage{
lensub(x)
}
\arguments{
\item{x}{a \code{regperiod_range}}
}
\value{
The number of subperiods in the range, or \code{Inf} is the
range is not bounded
}
\description{
Returns the number of subperiods in a \link{regperiod_range} object.
}

