% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movav.R
\name{movav}
\alias{movav}
\alias{movavb}
\alias{movavc}
\title{Moving average of a timeseries}
\usage{
movavb(x, order, keep_range = TRUE)

movavc(x, order, keep_range = TRUE)
}
\arguments{
\item{x}{a \code{\link[stats]{ts}} or \code{\link{regts}} object}

\item{order}{the order of the moving average}

\item{keep_range}{If \code{TRUE} (the default), then  the output
timeseries has the same period range as the input timeseries.
Then the result timeseries will have \code{order} NA values. For
\code{movavc} these NAs will apear on the left side of and for \code{movavc}
they will be distributed over both sides.
If \code{TRUE} then the result timeseries is \code{order} periods
shorter than the input timeseries.}
}
\value{
a \code{regts} object with the moving average values
}
\description{
Function \code{movavb} computes the backward moving avergage and
function \code{movavc} the centered moving average.

For example, the backward moving average of order 3 is defined as

\eqn{A[t] = (x[t-2] + x[t-1] + x[t]) / 3},

while the centered moving average of order 3 is calculated as

\eqn{A[t] = (x[t - 1] + x[t] + x[t + 1]) / 3}

Currently, the centered moving average has only been implemented for
odd orders.
}
\section{Functions}{
\itemize{
\item \code{movavb}: Backward moving average

\item \code{movavc}: Centered moving average (currently only for odd orders)
}}

\examples{
x <- regts(rnorm(10), start = "2018Q1")

movavb(x, order = 3)

movavc(x, order = 3, keep_range = FALSE)
}
