% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_utils.R
\name{na_skip}
\alias{na_skip}
\title{Function for removing leading and trailing NAs}
\usage{
na_skip(x, sides = c("both", "left", "right"), is.na = c("all", "any"))
}
\arguments{
\item{x}{a regts object}

\item{sides}{character with values \code{"both",} \code{"left"} or
\code{"right"} to remove NAs from
both sides (the default), just from the left side or just from the right side.}

\item{is.na}{character with values \code{"all"} or \code{"any"}. If \code{"all"}
(default) then a row will be regarded as \code{NA} only if all elements in
the row are \code{NA}. If \code{"any"} then a row will be regarded as \code{NA}
if it has any \code{NA}s. For one dimensional regts objects this argument
has no effect.}
}
\value{
A \code{regts} object in which leading and/or trailing NAs have been removed.
}
\description{
This function removes leading or trailing Nas or both from a (multivariate)
regts object.
For multivariate regts a row will default be regarded as NA if all elements
in the row are NA.  Specify \code{"any"} to change this behaviour.
}
\examples{
# remove only leading NAs
ts1 <- regts(c(NA,1,3,NA,4,8,NA), start = "2000")
na_skip(ts1, sides = "left")

# remove trailing NAs if all elements in the row are NA
data <- matrix(c(1,3,NA,2,5,NA,3,7,NA), ncol = 3)
rts <- regts(data, start = "2010Q2", names = c("a", "b", "c"))
na_skip(rts, sides = "right")

data <- matrix(c(NA,3,NA,NA,5,6,NA,7,NA), ncol = 3)
rts <- regts(data, start = "2010Q1", names = c("a", "b", "c"))
# remove leading/trailing NAs if all elements in the row are NA
na_skip(rts)
# or remove NAs if any NA occurs in that row
na_skip(rts, is.na = "any")
}

