% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_trim.R
\name{na_trim}
\alias{na_trim}
\title{Function for removing leading and trailing NAs}
\usage{
na_trim(x, method = c("both", "first", "last"), is_na = c("all", "any"))
}
\arguments{
\item{x}{a \code{\link{regts}} object}

\item{method}{character strings with values \code{"both",} \code{"first"} or
\code{"last"} to remove NAs at both ends (by default), just at the start
or just at the end.}

\item{is_na}{character strings with values \code{"all"} or \code{"any"}.
If \code{"all"} #' (default) then a row will be regarded as \code{NA} only if
all elements in #' the row are \code{NA}. If \code{"any"} then a row will be
regarded as \code{NA} #' if it has any \code{NA}s. For one dimensional regts
objects this argument #' has no effect.}
}
\value{
A \code{\link{regts}} object in which leading and/or trailing NAs have
been removed.
}
\description{
This function removes leading or trailing NAs or both from a (multivariate)
regts object.
For multivariate regts a row will by default be regarded as NA if all elements
in the row are NA. Use argument \code{is.na = "any"} to change this behaviour.
}
\examples{
# remove only leading NAs
ts1 <- regts(c(NA,1,3,NA,4,8,NA), start = "2000")
na_trim(ts1, method = "first")

# remove trailing NAs
data <- matrix(c(1,3,NA,2,5,NA,3,7,NA), ncol = 3)
rts <- regts(data, start = "2010Q2", names = c("a", "b", "c"))
na_trim(rts, method = "last")

data <- matrix(c(NA,3,NA,NA,5,6,NA,7,NA), ncol = 3)
rts <- regts(data, start = "2010Q1", names = c("a", "b", "c"))
# remove leading NAs if all elements in the row are NA
na_trim(rts, method = "first")
# or remove rows if any NA occurs in that row
na_trim(rts, is_na = "any")
}
\seealso{
\code{\link{zero_trim}}
}
