% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_range.R
\name{nperiod}
\alias{nperiod}
\title{Return the number of periods in a period range}
\usage{
nperiod(x)
}
\arguments{
\item{x}{a \code{\link{period_range}} or an object that can be coerced
to a \code{period_range}}
}
\value{
The number of periods in the range, or \code{Inf} if the
range is not bounded
}
\description{
Return the number of periods in a period range
}
\examples{
range <- period_range("2010Q2", "2011Q3")
nperiod(range)  # the result will be 6
}
