% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_data_frame.R
\name{numeric_data_frame}
\alias{numeric_data_frame}
\title{Convert all non-numeric columns of a dataframe to numerical columns.}
\usage{
numeric_data_frame(x, dec = ".")
}
\arguments{
\item{x}{a data frame}

\item{dec}{decimal separator in number strings. The default is \code{"."}}
}
\description{
This function converts all non-numeric columns of a data frame to
numerical columns. Integer columns are left unchanged. If the data frame
contains character columns with texts that cannot be converted to
a numerical value, then this function gives a warning with a list
of texts that could not be converted.
}
\examples{
df <- data.frame(a = c("1.123", "x", NA), b = c("1", "", "john"),
                 c = 10 : 12)
numeric_data_frame(df)
}
