% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_matrix.R
\name{numeric_matrix}
\alias{numeric_matrix}
\title{Convert a data frame to a numeric matrix.}
\usage{
numeric_matrix(x, dec = ".")
}
\arguments{
\item{x}{a data frame}

\item{dec}{decimal separator in number strings. The default is \code{"."}}
}
\description{
Returns the matrix obtained by converting all the variables in a data frame
to numeric mode and then binding them together as the columns of a matrix.
Factors and ordered factors are replaced by their internal codes.
of texts that could not be converted. This implementation is
significantly faster than the function \code{\link[base]{data.matrix}}
of the base package.
}
\examples{
df <- data.frame(a = c("1.123", "x", NA), b = c("1", "", "john"),
                 c = 10 : 12)
numeric_matrix(df)
}
