% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_columns.R
\name{order_columns}
\alias{order_columns}
\title{Order the columns of a matrix, data frame or timeseries alphabetically.}
\usage{
order_columns(x)
}
\arguments{
\item{x}{An R object with column names (a \code{matrix}, \code{ts}, \code{regts}
or \code{data.frame}).}
}
\value{
The same object as \code{data}, but with alphabetically sorted columns.
}
\description{
This function orders the columns of a matrix, data frame or
multivariate timeseries alphabetically.
}
\examples{
# multivariate timeseries
ts1 <- regts(matrix(1:9, ncol = 3), start = "2010Q4",
             names = c("c", "b", "a"))
print(order_columns(ts1))
}
\seealso{
\code{\link{rename_cols}}
}
