% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{period}
\alias{period}
\title{Create a \code{\link{period}} object based on a character string or numeric.}
\usage{
period(x, frequency = NA)
}
\arguments{
\item{x}{a character string or numeric}

\item{frequency}{frequency of the period object.
Argument \code{frequency} is mandatory if a general period format
such as \code{"2011-1"} has been specified}
}
\value{
a \code{period} object
}
\description{
Possible character string formats are for example \code{"2010Q2"},
\code{"2010M2"}, \code{"2011"} or \code{"2011-1"}.
The period may be prefixed with a \code{"T"} or \code{"Y"}.
Possible numeric formats are for example \code{2011} or \code{2011.25}
(the second quarter of 2011).
}
\examples{
period("2010Q3")
period("2010-2", frequency = 3)
period(2015)
period(2010.25, frequency = 4)
}
\seealso{
\code{\link{as.period}} and \code{\link{period_range}}
}
