% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_range.R
\name{period_range}
\alias{period_range}
\alias{as.period_range}
\title{Create a \code{\link{period_range}} object.}
\usage{
period_range(p1, p2 = p1, frequency = NA)

as.period_range(x, frequency = NA, ...)
}
\arguments{
\item{p1}{the first period (a \code{period}, an object that can be coerced
to a \code{period}, or \code{NULL}). If \code{p1} is \code{NULL} the lower
bound of the period range is undetermined. \code{p1} can also be
a character string specifying a period range, for example
\code{"2010Q2/2011Q3"}).}

\item{p2}{the last period (a \code{period}, an object that can be coerced
to a \code{period}, or \code{NULL}).
If \code{p2} is \code{NULL}, the upper bound of the period range is undetermined.}

\item{frequency}{frequency of the period objects. This argument is mandatory
if argument \code{p1} or \code{p2} is a character with general period format
without frequency indicator (e.g. \code{"2011-1"})}

\item{x}{an R object}

\item{...}{additional arguments to be passed to or from methods (currently
not used in package \code{regts})}
}
\value{
a \code{period_range} object
}
\description{
A \code{period_range} object represents an interval of periods,
for example a period from \code{"2017Q2"} to \code{"2019Q3"}.
Function \code{period_range}  creates a \code{period_range} from
a single character string (e.g. \code{"2017Q2/2019Q3"}, see Details)
or from two R objects that can be coerced to \code{period} objects.
\cr\cr
Function \code{as.period_range} coerces an R object to a \code{period_range}
if possible.
}
\details{
It is possible to create a \code{period_range} from
a single string specifying a period range, for example
\code{"2017Q2/2019Q3"}. For this format, the first and last
period are separated by \code{"/"}. The first and last period
are specified according to the same format recognized by function
\code{\link{period}}. The first or last period may be omitted
(e.g. \code{"2017Q3/"})), in that case the period range
has no lower or upper bound.
}
\examples{
# two methods to create a period_range from 2010Q2 to 2016Q3
period_range("2010Q2", "2016Q3")
period_range("2010Q2/2016Q3")

# create a period_range for the first 5 quarters after 2013Q2
p1 <- period("2013Q3")
period_range(p1, p1 + 5)

# create a period_range from 2010Q2 with no upper bound
period_range("2010Q2", NULL)

# create a period_range for a timeseries with frequency 2 (half year)
period_range("2010-2", "2016-2", frequency = 2)

# convert a period object to a period_range with equal start and end period
p <- period("2010Q2")
as.period_range(p)
}
\seealso{
\code{\link{period}}, \code{\link{nperiod}},
\code{\link{start_period}} and \code{\link{end_period}}
}
