% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printobj}
\alias{printobj}
\title{printobj print the name, class and value of an object}
\usage{
printobj(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{further arguments passed to \code{print}}
}
\description{
This function prints the name, class and value of its argument.
The value is printed using the standard \code{\link{print}} functions.
It returns the value of the argument invisbly
}
\examples{
x <- regts(1:5, start = "2017Q2")
printobj(x)
printobj(2 * x)
}
