% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_range_utils.R
\name{range_intersect/range_union}
\alias{range_intersect/range_union}
\alias{range_intersect}
\alias{range_union}
\title{Calculate the intersection or union of two \code{\link{period_range}} objects.}
\usage{
range_intersect(range1, range2)

range_union(range1, range2)
}
\arguments{
\item{range1}{a \code{period_range} object or object that can be coerced to
a period_range}

\item{range2}{another such (\code{period_range}) object}
}
\value{
the intersection or union of \code{range1} and \code{range2}.
If there are no common periods \code{range_intersection} returns \code{NULL}
}
\description{
These functions calculate the intersection or union of two \code{\link{period_range}}
objects. The start and end periods of the period_range objects may
not be \code{NULL}.
}
\examples{
range1 <- period_range("2016Q1", "2017Q4")
range2 <- period_range("2017Q1", "2018Q2")
range_intersect <- range_intersect(range1, range2)
range_union <- range_union(range1, range2)

}
