% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_range_utils.R
\name{range_union}
\alias{range_union}
\title{Calculates the union of two period_range objects.}
\usage{
range_union(range1, range2)
}
\arguments{
\item{range1}{a \code{period_range} object}

\item{range2}{another \code{period_range} object}
}
\value{
the union of \code{range1} and \code{range2}.
}
\description{
This function calculates the union of two period_range objects
The start and end periods of the period_range objects may not be NULL.
}
\examples{
range1 <- period_range("2016Q1", "2017Q4")
range2 <- period_range("2017Q1", "2018Q2")
range_union <- range_union(range1, range2)
}
\seealso{
\code{\link{regrange_intersect}}
}
