% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod.R
\name{regperiod}
\alias{regperiod}
\title{Create a \code{regperiod} object based on a string or an integer.}
\usage{
regperiod(x, frequency = NA)
}
\arguments{
\item{x}{a single string or integer}

\item{frequency}{frequency of the regperiod object.
Argument \code{frequency} is mandatory if a general regperiod format
such as "2011-1" has been specified}
}
\value{
a \code{regperiod} object
}
\description{
Possible string formats are for example \code{"2010Q2"},
\code{"2010M2"}, \code{"2011"} or \code{"2011-1"}.
}
\examples{
regperiod("2010Q3")
regperiod("2010-4", frequency = 3)
regperiod(2015)
}

