% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod_range.R
\name{regrange_intersect}
\alias{regrange_intersect}
\title{Calculates the intersection of two \code{regperiod_range} objects.}
\usage{
regrange_intersect(range1, range2)
}
\arguments{
\item{range1}{a \code{regperiod_range} object}

\item{range2}{another \code{regperiod_range} object}
}
\value{
the intersection of \code{range1} and \code{range2}. Returns
\code{NULL} if there is no intersection
}
\description{
Calculates the intersection of two \code{regperiod_range} objects.
}

