% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{regts.intersect}
\alias{cbind.regts}
\alias{regts.intersect}
\alias{regts.union}
\title{Bind Two or More Timeseries}
\usage{
regts.intersect(..., dframe = FALSE)

regts.union(..., dframe = FALSE)

cbind.regts(...)
}
\arguments{
\item{...}{two or more univariate or multivariate time series,
or objects which can coerced to time series}

\item{dframe}{logical if \code{TRUE} return the result as a data frame}
}
\value{
A \code{regts} timeseries object is \code{dframe} is
\code{FALSE}, otherwise a data frame.
}
\description{
Bind time series which have a common frequency using the methods
\code{\link[stats]{ts.intersect}} and \code{\link[stats]{ts.union}} of the
\code{stats} package. The implementation for \code{regts} converts the value
returned by \code{ts.intersect} and \code{ts.union}
to a \code{regts} and also takes care of the timeseries labels
(if present). The \code{cbind} methods for \code{regts} objects works
as \code{regts.union}, except that the argument \code{dframe} is
not used.
}

