% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{rel2index}
\alias{rel2index}
\title{Calculates an index timeseries from a timeseries with relative mutations}
\usage{
rel2index(x, base_period = start_period(x), scale = 1)
}
\arguments{
\item{x}{a \code{\link[stats]{ts}} or \code{\link{regts}}}

\item{base_period}{base period of the index timeseries}

\item{scale}{the value of the index series at the base period}
}
\description{
Calculates an index timeseries from a timeseries with relative mutations
}
