% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_na_columns.R
\name{remove_na_columns}
\alias{remove_na_columns}
\title{Function for removing columns with NAs in a (multivariate) regts}
\usage{
remove_na_columns(x, is_na = c("all", "any"))
}
\arguments{
\item{x}{a regts object}

\item{is_na}{character with values \code{"all"} or \code{"any"}.
If \code{"all"} (default) then a column will be regarded as \code{NA} only
if all elements in the column are \code{NA}. If \code{"any"} then a column
will be regarded as \code{NA} if it has any \code{NA}s.}
}
\value{
A \code{regts} object in which NA columns have been removed.
}
\description{
This function removes columns with NA values from a (multivariate)
\code{\link{regts}}.
A column will by default be regarded as NA if all elements
in the column are NA.  Specify argument \code{is.na = "any"} to change
this behaviour. If all columns are removed the function returns \code{NULL}.
}
\examples{
# remove columns with all NAs
data <- matrix(c(1,3, 5, NA, NA, NA,3,7,9), ncol = 3)
rts <- regts(data, start = "2010Q2", names = c("a", "b", "c"))
remove_na_columns(rts)

data <- matrix(c(NA,3,NA,NA,5,6,NA,7,9), ncol = 3)
rts <- regts(data, start = "2010Q1", names = c("a", "b", "c"))
remove_na_columns(rts, is_na = "any")
}
