% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_utils.R
\name{rename_cols}
\alias{rename_cols}
\title{Rename the columns of a matrix or timeseries.}
\usage{
rename_cols(.data, ...)
}
\arguments{
\item{.data}{An R object with column names (a \code{matrix}, \code{ts}, \code{regts}
or \code{data.frame})}

\item{...}{Use \code{new_name = old_name} to rename selected
variables. Alternatively, specify a named character vector (see Examples).}
}
\value{
An object with the same type and contents as \code{.data}, except
that the columns are renamed.
}
\description{
Function \code{rename_cols} renames the columns of a matrix using the
\code{new_name = old_name} syntax.
This works similarly as functions \code{\link[dplyr]{rename}} of the
\code{dplyr} package, but also works
on matrix objects (a \code{matrix} or \code{ts}) objects.
}
\examples{
data <- regts(matrix(1:20, ncol = 4), start = "2010Q2",
              names = c("a1", "a2", "b1", "b1"))

# Use new_name = old_name syntax
rename_cols(data, x = a1, y = b1)

# Use named character vector
rename_cols(data, c(x = "a1", y = "b1"))

# Rename the first column (new column name is 'p'.)
rename_cols(data, p = 1)
}
\seealso{
\code{\link{tag_colnames}} and \code{\link{change_colnames}}
}
