% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_end_period.R
\name{start_period/end_period}
\alias{end_period}
\alias{end_period.regperiod_range}
\alias{end_period.ts}
\alias{start_period}
\alias{start_period.regperiod_range}
\alias{start_period.ts}
\alias{start_period/end_period}
\title{Returns the start or end period of a timeseries object or a
\code{\link{regperiod_range}}}
\usage{
start_period(x)

end_period(x)

\method{start_period}{regperiod_range}(x)

\method{start_period}{ts}(x)

\method{end_period}{regperiod_range}(x)

\method{end_period}{ts}(x)
}
\arguments{
\item{x}{a \code{regts} or \code{regperiod_range} object}
}
\value{
A \code{regperiod} object representing the first or last period of 
the range. The return value can be \code{NULL} if argument \code{x} is a
\code{regperiod_range} with no lower or upper boundary.
}
\description{
This function returns the start or end period of a
timeseries object (a \code{\link{regts}} or \code{\link[stats]{ts}}))
or a \code{regperiod_range}.
}

