% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_end_period.R
\name{start_period}
\alias{end_period.regperiod_range}
\alias{end_period.ts}
\alias{start_period}
\alias{start_period.regperiod_range}
\alias{start_period.ts}
\title{Returns the start period of a \link{regts} or a \link{regperiod_range}}
\usage{
start_period(x)

\method{start_period}{regperiod_range}(x)

\method{start_period}{ts}(x)

\method{end_period}{regperiod_range}(x)

\method{end_period}{ts}(x)
}
\arguments{
\item{x}{a \link{regts} or \link{regperiod_range{ object}}}
}
\value{
a \link{regperiod} object representing the first period of the
range. The return value can be \code{NULL} if argument \code{x} is a
\code{regperiod_range} with no lower boundary.
}
\description{
Returns the start period of a \link{regts} or a \link{regperiod_range}
}
\section{Methods (by class)}{
\itemize{
\item \code{regperiod_range}: Returns the first period

\item \code{ts}: Returns the first period

\item \code{regperiod_range}: Returns the end period

\item \code{ts}: Returns the end period
}}

