% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_utils.R
\name{tag_colnames}
\alias{tag_colnames}
\title{Append a tag to the column names of a matrix or data frame}
\usage{
tag_colnames(x, tag)
}
\arguments{
\item{x}{an R object with column names (e.g. a \code{data.frame},
\code{matrix}, \code{ts} or \code{regts})}

\item{tag}{a character}
}
\value{
an R object with modified columm names
}
\description{
This function adds a suffix to all column names
of an R object.
}
\examples{
data <- regts(matrix(1:8, ncol = 2), start = "2010Q2", names = c("a", "b"))

#  append the column names with a tag "_input"
tag_colnames(data, "_input")
}
\seealso{
\code{\link{change_colnames}}
}
