% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_tail.R
\name{tail.regts}
\alias{tail.regts}
\title{Printing tail of a regts}
\usage{
\method{tail}{regts}(x, method = c("first", "last"), size = 10, ...)
}
\arguments{
\item{method}{if \code{first} then the first part of the timeseries is printed,
otherwise (\code{last}) the last part of the timeseries is printed.}

\item{size}{is the number of timeseries that are printed. Default = 10.}

\item{...}{arguments passed to \code{\link{tail}}.}
}
\description{
With this function the tail of a regts can be printed. The generic function
\code{\link{tail}} is extended with parameters \code{method} and \code{size}.
}
\examples{
data <- regts(matrix(1:200, ncol = 20), start = "2010Q2",
              names = paste0("abc", 1:20))
tail(data)
tail(data, method = "last", size = 5)
}
\seealso{
\code{\link{head.regts}}
}
