% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_tail.R
\name{topleft}
\alias{topleft}
\title{Returns the topleft part of a regts.}
\usage{
topleft(x, n = 6L, ncol = 10L, ...)
}
\arguments{
\item{x}{a multivariate \code{\link{regts}}}

\item{n}{a single integer. Length period for the resulting object.}

\item{ncol}{a single integer. Number of columns in \code{\link{regts}}.
By default only the first 10 columns are printed.}
}
\description{
Returns the topleft part of a regts.
}
\examples{
data <- regts(matrix(1:200, ncol = 20), start = "2010Q2",
              names = paste0("abc", 1:20))
topleft(data)
}
\seealso{
\code{\link{head}}, \code{\link{tail}}
}
