% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_df.R
\name{transpose_df}
\alias{transpose_df}
\title{Transpose a \link{data.frame}}
\usage{
transpose_df(x, colname_column, label_column)
}
\arguments{
\item{x}{a data frame}

\item{colname_column}{the name or the index of the column that will be used
as the column names of the transposed data frame. By default the
row names of the original data frame are used as column names of the new data
 frame.}

\item{label_column}{the name or the index of the column that will be used
to create labels for the transposed data frame. By default no labels
are created}
}
\value{
the transposed data frame
}
\description{
The function transposes a data frame. If the data frame contains
labels that has been set by the function \link{label} of the
package \code{Hmisc}, then the first column of the returned data frame will
contain the labels. Conversely, you can specify the column that will become
the labels of the transposed data frame.
}
\examples{
df <- data.frame(variables = c("a", "b"), labels = c("Variabele a",
                 "Variable c"), x = 1:3, y = 10:12)
df_t <- transpose_df(df, colname_column = 1, label_column = 2)
print(df_t)
print(transpose_df(df_t))
}

