% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_labels.R
\name{ts_labels}
\alias{ts_labels}
\alias{ts_labels<-}
\title{Timeseries labels}
\usage{
ts_labels(x)

ts_labels(x) <- value
}
\arguments{
\item{x}{a \code{\link{regts}}}

\item{value}{a character vector with the labels or \code{NULL}. The length
should be equal to the number of columns. Specify \code{NULL} to remove all
labels.}
}
\description{
Retrieve or set labels for the timeseries. Timeseries labels
can be used to give a description of the contents of the timeseries.
The labels are stored in a named list: the names are the timeseries names
(column names), and the values the corresponding label.
}
\section{Functions}{
\itemize{
\item \code{ts_labels}: Retrieve timeseries labels

\item \code{ts_labels<-}: Sets the timeseries labels
}}

\examples{
ts <- regts(matrix(1:6, ncol = 2), start = "2016Q2", names = c("a", "b"))
ts_labels(ts) <- c("Timeseries a", "Timeseries b")
print(ts_labels(ts))

# print the column names and labels as a nice data.frame
print(as.data.frame(ts_labels(ts)))
}
\seealso{
\code{\link{regts}}, \code{\link{update_ts_labels}}
}
