% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdif.R
\name{tsdif}
\alias{tsdif}
\title{Calculates the difference between two multivariate ts objects}
\usage{
tsdif(x1, x2, tol = 0, fun = function(x1, x2) abs(x1 - x2))
}
\arguments{
\item{x1}{the first timeseries (a \link{regts} or \link{mts} object).}

\item{x2}{the second timeseries (a \link{regts} or \link{mts} object).}

\item{tol}{difference tolerance (by default zero). Differences smaller
than tol are ignored.}

\item{fun}{function to compute differences. This function should accept
two arguments (two numbers) for which the difference is computed.
By default the absolute difference is computed. A useful function for
computing difference is \link{cvgdif}, which computes relative differences
if the absolute value of \code{x2} is larger than 1.}
}
\value{
a list with the following components
 \item{tol}{The tolerance parameter}
 \item{missing_names1}{The names of columns present in \code{x2} but missing
                       in \code{x1}}
 \item{missing_names2}{The names of columns present in \code{x1} but missing
                       in \code{x2}}
 \item{common_names}{The names of the common columns of \code{x1} and
                    \code{x2}}
 \item{has_dif}{\code{TRUE} if any difference larger than \code{tol} was
                found}
 \item{difnames}{The names of the common columns with differences
                  larger than \code{tol}}
 \item{dif}{The computed differences for the common columns with
            differences are larger than \code{tol}, or \code{NULL} is there
            are no differences}
}
\description{
This function calculates the difference between two multivariate timeseries
objects \code{x1} and \code{x2} for the common columns.
The two timeseries must have the same frequency, but may have a different
period ranges. The difference is computed for the intersection of the two
period ranges (for the non-overlapping periods the result will be
\code{NA}). Two \code{NA} values are considered to be equal.
}
\examples{
library(regts)

# create two timeseries x1 and x2
x1 <- regts(matrix(data = rep(1:27), nc = 3), start = "2008Q4",
           names = c("a", "b", "c"))
x2 <- x1 + 0.001
colnames(x2) <- c("a", "b", "d")

dif1 <- tsdif(x1, x2)
print(dif1)

# use the function regts::cvgdif (convergence difference)
dif2 <- tsdif(x1, x2, fun = cvgdif)
print(dif2)

# ignore differences smaller than 1e-4
dif3 <- tsdif(x1, x2, tol = 1e-4, fun = cvgdif)
print(dif3)


}

