% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_regts.R
\name{tsupdate}
\alias{tsupdate}
\title{Updates a multivariate timeseries with another multivariate timeseries
objects}
\usage{
tsupdate(x1, x2, method = c("tsupd", "tsupdna", "replace", "tsupdval"))
}
\arguments{
\item{x1}{the first timeseries (a multivariate \code{\link{regts}} or
\code{\link[stats]{ts}} object).}

\item{x2}{the second timeseries (a multivariate \code{regts} or \code{ts} object).}

\item{tol}{difference tolerance (by default zero). Differences smaller
than tol are ignored.}

\item{fun}{function to compute differences. This function should accept
two arguments (two numbers) for which the difference is computed.
By default the absolute difference is computed. A useful function for
computing differences is \code{\link{cvgdif}}, which computes relative differences
if the absolute value of \code{x2} is larger than 1.}
}
\value{
a list with the following components
}
\description{
This function can be used to update a multivariate regts object with
another multivariate regts object.
The result is the updated multivariate regts.
}
\details{
This function calculates the difference between common columns
of two multivariate timeseries objects \code{x1} and \code{x2}.
The two timeseries must have the same frequency, but may have a different
period range. The difference is computed for the union of the two
period ranges (for the non-overlapping periods the result will be
\code{NA}). Two \code{NA} values are considered to be equal.
}
\examples{
library(regts)

# create two timeseries x1 and x2
x1 <- regts(matrix(data = rep(1), nc = 3), period = "2000/2002",
            names = c("a", "b", "c"))
x2 <- regts(matrix(data = rep(2), nc = 3), period = "2000/2001",
            names = c("a", "c", "d"))
x1 <- regts(matrix(data = rep(1:127), nc = 3), start = "2000",
           names = c("a", "b", "c"))
x2 <- x1 + 0.001
colnames(x2) <- c("a", "b", "d")

}
\seealso{
\code{\link{regts}}
}
