% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ts.R
\name{update_ts}
\alias{update_ts}
\title{Update a multivariate timeseries with another multivariate timeseries
object}
\usage{
update_ts(x1, x2, method = c("upd", "updna", "updval", "replace"))
}
\arguments{
\item{x1}{the first timeseries (a multivariate \code{\link{regts}} or
\code{\link[stats]{ts}} object).}

\item{x2}{the second timeseries (a multivariate \code{regts} or \code{ts} object).}

\item{method}{four different ways to update the timeseries.
By default the timeseries are updated. This behaviour can be changed by
using one of the other methods. See details.}
}
\value{
an updated multivariate \code{\link{regts}} object.
}
\description{
This function can be used to update a multivariate (reg)ts object with
another multivariate (reg)ts object.
The result is an updated multivariate \code{\link{regts}} object.
}
\details{
The timeseries can be updated in four different ways:

\code{upd} the first timeseries are updated with the second timeseries.
The two timeseries must have the same frequency, but may have a different
period range. The update is computed for the union of the two period ranges
(for the non-overlapping periods the result will be \code{NA}).

\code{updna} if method \code{updna} is selected instead of \code{upd},
only NA values in the first timeseries will be updated

\code{updval} if method \code{updval} is selected instead of \code{upd},
the values in the first timeseries are only replaced with valid (i.e. non-NA)
values from the second timeseries.

\code{replace} if method \code{replace} is selected the first
timeseries is replaced by the second timeseries

The first three update methods are only employed at the common columns in
both timeseries.
The non overlapping columns in both timeseries are added to the result.
}
\examples{
library(regts)

x1 <- regts(matrix(data = rep(1:9), nc = 3), period = "2000/2002",
            names = c("a", "b", "c"))
x2 <- regts(matrix(data = rep(10:15), nc = 3), period = "2000/2001",
            names = c("a", "c", "d"))
update_ts(x1, x2, method = "upd")

}
\seealso{
\code{\link{regts}}
}
