% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_trim.R
\name{zero_trim}
\alias{zero_trim}
\title{Function for removing leading and trailing zeros}
\usage{
zero_trim(x, method = c("both", "first", "last"))
}
\arguments{
\item{x}{a regts object}

\item{method}{character with values \code{"both",} \code{"first"} or
\code{"last"} to remove ze5ros at both ends (by default), just at the start
or just at the end.}
}
\value{
A \code{regts} object in which leading and/or trailing zeros have been removed.
}
\description{
This function removes leading or trailing zeros or both from a (multivariate)
regts object.
For multivariate regts a row will by default be regarded as 0 if all elements
in the row are 0.  Specify \code{"any"} to change this behaviour.
}
\examples{
# remove only leading zeros
ts1 <- regts(c(0, 1, 3, 0, 4, 8, 0), start = "2000")
zero_trim(ts1, method = "first")

# remove trailing zeros
data <- matrix(c(1, 3, 0,2, 5, 0, 3, 7, 0), ncol = 3)
rts <- regts(data, start = "2010Q2", names = c("a", "b", "c"))
zero_trim(rts, method = "last")

data <- matrix(c(0, 3, 0, 0, 5, 6, 0, 7, 0), ncol = 3)
rts <- regts(data, start = "2010Q1", names = c("a", "b", "c"))
# remove leading NAs if all elements in the row are NA
zero_trim(rts, method = "first")
}
