% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_gr.R
\name{aggregate_gr}
\alias{aggregate_gr}
\title{Convert timeseries with absolute or relative growth rates to a lower
frequency}
\usage{
aggregate_gr(x, method = c("dif1s", "dif1", "pct", "rel"),
  nfrequency = 1)
}
\arguments{
\item{x}{a \code{\link[stats]{ts}} of \code{\link{regts}} object}

\item{method}{Aggregation method: \code{"dif1s"}, \code{"dif1"}, \code{"pct"}
or \code{"rel"}. See details.}

\item{nfrequency}{the frequency of the result. This should be higher than
the frequency of timeseries \code{x}}
}
\value{
a \code{regts} with frequency \code{nfrequency}
}
\description{
Special aggregation methods are needed for timeseries that contain absolute,
relative or percentage changes, the 'cumulative growth methods'. There are
four different type of methods for different types of input timeseries.
}
\details{
The \code{dif1s} and \code{dif} methods assume that the input timeseries
contain a first difference of length 1 (for \code{dif1s} the input is also
scaled). The result is a first difference of length 1 in the output frequency.
The \code{pct} and \code{rel} methods assume timeseries that contain
percentage or relative change. They calculate the exact percentage or
relative change for the output timeseries.

More details of the aggregation methods are  provided in the Vignette
"Introduction to package regts".
}
\examples{
ts_q <- regts(rnorm(10), start = "2016Q1")
aggregate_gr(ts_q, method = "dif1s")

ts_m <- regts(matrix(rnorm(20), ncol = 2), start = "2017M1", names = c("a", "b"))
aggregate_gr(ts_m, method = "rel", nfrequency = 4)
}
