% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_gr.R
\name{aggregate_gr}
\alias{aggregate_gr}
\title{Converts timeseries with absolute or relative growth rates to a lower
frequency}
\usage{
aggregate_gr(x, frequency = 1)
}
\arguments{
\item{x}{a \link{ts} of \link{regts} object}

\item{frequency}{the frequency of the result, should be higher than
the frequency of timeseries \code{x}}
}
\value{
a \link{regts} with lower frequency
}
\description{
Converts timeseries with absolute or relative growth rates to a lower
frequency
}

