% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_gr.R
\name{aggregate_gr}
\alias{aggregate_gr}
\title{Converts timeseries with absolute or relative growth rates to a lower
frequency}
\usage{
aggregate_gr(x, method, nfrequency = 1)
}
\arguments{
\item{x}{a \code{\link[stats]{ts}} of \code{\link{regts}} object}

\item{method}{Aggregation method: \code{"cgr"}, \code{"cgrs"}, \code{"cgru"}
or \code{"cgrc"}. Consult the Isis reference manual for an explanation of
these methods}

\item{nfrequency}{the frequency of the result, should be higher than
the frequency of timeseries \code{x}}
}
\value{
a \code{regts} with frequency \code{frequency}
}
\description{
Converts timeseries with absolute or relative growth rates to a lower
frequency
}

