% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame.regts}
\title{Convert a \code{\link{regts}} to a \code{\link[base]{data.frame}}}
\usage{
\method{as.data.frame}{regts}(x, ..., rowwise = FALSE, row_names = TRUE,
  period_as_Date = FALSE)
}
\arguments{
\item{x}{a \code{\link{regts}}}

\item{...}{additional arguments to be passed to methods.}

\item{rowwise}{a logical value: should the timeseries be stored rowwise
or columnwise in the data frame? Defaults to \code{FALSE}}

\item{row_names}{Whether to create row names. If \code{FALSE},
then an additional column with name \code{"period"} or \code{"name"} is
created for columnwise or rowwise timeseries, respectively.}

\item{period_as_Date}{A logical (default \code{FALSE}).
If \code{TRUE} the periods are stored as \code{\link[base]{Date}} objects.
If you use this option it is recommended to use \code{row_names = FALSE},
since row names are always characters. If row names are used
the \code{Date} objects are converted to characters,
using the  standard date format format \code{"\%Y-\%m-\%d"}
(see the documentation of function \code{\link[base]{strptime}}
for more information about date formats).}
}
\value{
A \code{\link[base]{data.frame}}
}
\description{
Convert a \code{\link{regts}} to a \code{\link[base]{data.frame}}
}
\details{
If the \code{regts} has labels and argument \code{rowwise} is \code{FALSE},
then the labels are added to columns of the data frame
These labels are visible in the data viewer.
}
\examples{
ts <- regts(matrix(1:4, ncol = 2) , start = "2015Q3", names = c("a", "b"),
           labels = c("Timeseries a", "Timeseries b"))
print(as.data.frame(ts))
}
