% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{as.data.frame.regts}
\alias{as.data.frame.regts}
\title{Conversion between dataframes and regts.}
\usage{
\method{as.data.frame}{regts}(x, ...)
}
\description{
Zoo has excellent facilities for this,
therefore we employ zoo for the time being.
Convert a regts to a dataframe.
For the time being use zoo as intermediate class.
}

