% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{as.data.frame.regts}
\alias{as.data.frame.regts}
\title{Convert a \code{\link{regts}} to a \code{\link[base]{data.frame}}}
\usage{
\method{as.data.frame}{regts}(x, ...)
}
\arguments{
\item{x}{a \code{\link{regts}}}

\item{...}{additional arguments to be passed to methods.}
}
\value{
A \code{\link[base]{data.frame}}
}
\description{
Convert a \code{\link{regts}} to a \code{\link[base]{data.frame}}
}
\details{
Unlike \code{\link{as.regts.ts}} this function adds row names with a period
indicator.
If the \code{regts} has labels, then the labels are added to the result
dataframe using the function \code{\link[Hmisc]{label}}
of package \code{Hmisc}.
}
\examples{
ts <- regts(matrix(1:4, ncol = 2) , start = "2015Q3", names = c("a", "b"),
           labels = c("Timeseries a", "Timeseries b"))
print(as.data.frame(ts))
}
