% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{as.period}
\alias{as.period}
\alias{as.period.character}
\alias{as.period.numeric}
\title{Coerce an R object to a period}
\usage{
as.period(x, ...)

\method{as.period}{character}(x, frequency = NA, ...)

\method{as.period}{numeric}(x, frequency = NA, ...)
}
\arguments{
\item{x}{any R object}

\item{...}{object passed to methods}

\item{frequency}{the frequency of the period. This argument is mandatory
if the frequency cannot be detected automatically.}
}
\value{
a \link{period}
}
\description{
Coerce an R object to a period
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Coerce a character string to a \code{period}

\item \code{numeric}: Coerce a numerical value string to a \code{period}
}}

\examples{
as.period("2010Q3")
as.period(2010)
}
