% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod_range.R
\name{as.period_range}
\alias{as.period_range}
\alias{as.period_range.character}
\alias{as.period_range.period}
\title{Coerce an R object to a \code{\link{period_range}}}
\usage{
as.period_range(x, frequency = NA, ...)

\method{as.period_range}{character}(x, frequency = NA, ...)

\method{as.period_range}{period}(x, ...)
}
\arguments{
\item{x}{any R object}

\item{frequency}{the frequency (mandatory for
\code{as.period_range.character}
if a period format without frequency indicator has been used,
e.g. \code{"2011-3"})}

\item{...}{arguments passed to methods (not used in this package)}
}
\value{
a \code{period_range}
}
\description{
Coerce an R object to a \code{\link{period_range}}
}
\details{
\code{as.period_range.character} converts a character string to a
period_range object. The format for the first and last period of
the range are specied as in \code{\link{period}}. The two periods strings
are separated by \code{"/"} (e.g. \code{"2010Q2/2016Q2"}).
The first or last period may be omitted: in that case the period range
has no lower of upper bound (e.g. \code{"2012Q3/"}). The \code{"/"}
separator is also optional: if omitted, then the period range
contains a single period.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Convert a character string to a
period_range object

\item \code{period}: Convert a single \code{\link{period}} to a
\link{period_range}.
}}

\examples{
as.period_range("2010Q2/2016Q3")
as.period_range("2010Q2/")
as.period_range("/2016Q3")
as.period_range("2016Q1")

# convert a period object to a period_range:
p <- period("2010Q2")
as.period_range(p)
}
