% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod.R
\name{as.regperiod}
\alias{as.regperiod}
\title{Coerce an R object to a regperiod}
\usage{
as.regperiod(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{object passed to methods}

\item{frequency}{the frequency of the regperiod. This is mandatory
if the frequency cannot be detected automatically.}
}
\value{
a \link{regperiod}
}
\description{
Coerce an R object to a regperiod
}

