% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regts.R
\name{as.regts}
\alias{as.regts}
\alias{as.regts.ts}
\alias{as.regts.data.frame}
\alias{as.regts.default}
\title{Coerce an object to a \code{\link{regts}} timeseries object}
\usage{
as.regts(x, ...)

\method{as.regts}{ts}(x, ...)

\method{as.regts}{data.frame}(x, time_column = 0, numeric = TRUE,
  fun = regperiod, ...)

\method{as.regts}{default}(x, ...)
}
\arguments{
\item{x}{an arbitrary R object}

\item{...}{arguments passed to \code{fun}}

\item{time_column}{the column names or numbers of the data frame
in which the time is stored. Specify \code{0} if the index is in the row names
of the data frame}

\item{numeric}{logical: should non numeric values be converted to numeric data.
By default they are converted to numeric. This can be changed by setting
\code{numeric = FALSE}.}

\item{fun}{a function for converting values in the time column to
\code{\link{regperiod}} objects}
}
\value{
a \code{regts} object
}
\description{
Coerce an object to a \code{\link{regts}} timeseries object
}
\section{Methods (by class)}{
\itemize{
\item \code{ts}: Coerce a \code{\link[stats]{ts}} to a
\code{\link{regts}}

\item \code{data.frame}: Convert a \code{\link{data.frame}} to a
\code{\link{regts}}

\item \code{default}: Default method to convert an R object to a
\code{\link{regts}}. This method first employs \code{\link[stats]{as.ts}}
and then \code{\link{as.regts.ts}}
}}

\examples{
# convert a ts to regts
x <- ts(1:3, start = c(2015,3), frequency = 4)
x <- as.regts(x)

# Now three examples for converting a data.frame

# create a data frame with timeseries and with the
# time index in the rownames, and convert to a regts
df <- data.frame(a = 1:3)
rownames(df) <- c("2015Q3", "2015Q4", "2016Q1")
ts <- as.regts(df)

# create a data frame with the time index in the first column and special
# time format "2015 3" instead of "2015Q3", and convert to regts
df <- data.frame(periods = c("2015 3", "2015 4", "2016 1"),  a = 1:3)
ts <- as.regts(df, time_column = 1, frequency = 4)

# create a dataframe with non numeric data and convert to regts
# Strings containing non numeric values are converted to NA
# Logical values TRUE/FALSE are converted to 1/0
df <- data.frame(a = c("1", "2", "X"), b = c(TRUE, FALSE, TRUE), stringsAsFactors = FALSE)
as.regts(df)
}
\seealso{
\code{\link{regts}}, \code{\link{is.regts}},
\code{\link{as.data.frame.regts}},
\code{\link{as.list.regts}}, \code{\link{start_period}}, \code{\link{end_period}}
}
