% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regts.R
\name{as.regts}
\alias{as.regts}
\alias{as.regts.data.frame}
\alias{as.regts.default}
\alias{as.regts.ts}
\title{Coerce an object to a \link{regts} timeseries object}
\usage{
as.regts(x, ...)

\method{as.regts}{ts}(x, ...)

\method{as.regts}{data.frame}(x, time_column = 0, fun = regperiod, ...)

\method{as.regts}{default}(x, ...)
}
\arguments{
\item{x}{an arbitrary R object}

\item{...}{arguments passed to \code{fun}}

\item{time_column}{the column names or numbers of the data frame
in which the time is stored. Specify \code{0} if the index is in the row names
of the data frame}

\item{fun}{a function for converting values in the time column to
\link{regperiod} objects}
}
\value{
a \link{regts} object
}
\description{
Coerce an object to a \link{regts} timeseries object
}
\section{Methods (by class)}{
\itemize{
\item \code{ts}: Coerce a \link{ts} to a \link{regts}

\item \code{data.frame}: Convert a \link{data.frame} to a \link{regts}

\item \code{default}: Default method to convert an R object to a \link{regts}.
This method #' first employs \link{at.ts} and then \link{as.regts.ts}
}}
\examples{
# convert a ts to regts
x <- ts(1:3, start = c(2015,3), frequency = 4)
x <- as.regts(x)

# Now two examples for converting a data.frame

# create a data frame with timeseries and with the
# time index in the rownames, and convert to a regts
df <- data.frame(a = 1:3)
rownames(df) <- c("2015Q3", "2015Q4", "2016Q1")
ts <- as.regts(df)

# create a data frame with the time index in the first column and special
time format "2015 3" instead of "2015Q3", and convert to regts
df <- data.frame(periods = c("2015 3", "2015 4", "2016 1"),  a = 1:3)
ts <- as.regts(df, time_column = 1, frequency = 4)
}
\seealso{
\link{regts}, \link{is.regts}, \link{as.data.frame.regts},
\link{as.list.regts}, \link{start_period}, \link{end_period}
}

