% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as_data_frame}
\alias{as_data_frame}
\alias{as_data_frame.regts}
\title{Convert a \code{\link{regts}} to a \code{\link[base]{data.frame}}}
\usage{
\method{as_data_frame}{regts}(x, ..., format = c("columnwise", "rowwise",
  "long"), period_as_date = FALSE, period_col = "period",
  name_col = "name", label_col = "label", value_col = "value")
}
\arguments{
\item{x}{a \code{\link{regts}} object.}

\item{...}{additional arguments to be passed to methods (not used)}

\item{format}{A character specifying the format (see Details): \code{"rowwise"},
\code{"columnwise"} or \code{"long"}. For the long format, \code{x} should not have
duplicate column names.}

\item{period_as_date}{A logical (default \code{FALSE}).
If \code{TRUE} the periods are stored as \code{\link[base]{Date}} objects.
Depending on arguments \code{rowwise} and \code{row_names}
the periods may appear in the row or column names of the result data frame.
In that case the dates are coerced to character vectors,
using the standard date format \code{"\%Y-\%m-\%d"}
(see the documentation of function \code{\link[base]{strptime}}
for more information about date formats).}

\item{period_col}{The name of the column with periods (default \code{"period"}),
used for the columnwise or long format.}

\item{name_col}{The name of the column with variable names
(default \code{"name"}), used for the rowwise or long format.}

\item{label_col}{The name of the column with labels of the variables
(default \code{"label"}) used for the rowwise or long format. If the timeseries
does not have labels then this argument is ignored.}

\item{value_col}{The name of the columns with values (default \code{"value"})
for the long format.}
}
\value{
A \code{\link[base]{data.frame}}
}
\description{
Convert a \code{\link{regts}} to a \code{\link[base]{data.frame}}
}
\details{
Three different format for the data frame are possbile.

\strong{1. columnwise (default)}

This is the default format. There is a column for each variable. The first
column (by default named \code{period}) will contain the periods, as shown below:

\if{html}{\out{<div class="sourceCode">}}\preformatted{period  a   b
2022    1  10
2023    2  20
}\if{html}{\out{</div>}}

If the \code{regts} has labels, then each column of the data frame gets an
attribute named \code{label} with the label as value. These labels are shown when
the data frame is opened in the Data Viewer of RStudio. The labels can
also be retrieved by using function \code{\link[labelled]{var_label}} from
package \code{labelled}.

\strong{2. rowwise}

For a rowwise data frame, there is a column for each period. For example,

\if{html}{\out{<div class="sourceCode">}}\preformatted{name   label       2022  2023
a      Variable a     1     2
b      Variable b    10    20
}\if{html}{\out{</div>}}

\strong{3. long format}

For a data frame with long format, there is one row for each observation.
For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ name  label        period value
 a     Variable a     2022     1
 a     Variable a     2023     2
 b     Variable b     2022    10
 b     Variable b     2023    20
}\if{html}{\out{</div>}}

Use argument \code{long = TRUE} to create such a data frame.
If the timeseries has no labels, the \code{label} column is missing.
}
\examples{
library(regts)
ts <- regts(matrix(1:4, ncol = 2) , start = "2015Q3", names = c("a", "b"),
           labels = c("Timeseries a", "Timeseries b"))
print(as_data_frame(ts))

print(as_data_frame(ts, format = "long"))
}
\seealso{
\code{\link{long_df_to_regts}}.
}
