% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_utils.R
\name{change_colnames}
\alias{change_colnames}
\title{Change the column names of a matrix or data frame by applying a function.}
\usage{
change_colnames(x, fun, ...)
}
\arguments{
\item{x}{an R object with column names (e.g. a \code{data.frame},
\code{matrix}, \code{ts} or \code{regts})}

\item{fun}{a function}

\item{...}{arguments passed to \code{fun}}
}
\value{
an R object with modified columm names
}
\description{
Change the column names of a matrix or data frame by applying a function.
}
\examples{
data <- regts(matrix(1:8, ncol = 2), start = "2010Q2", names = c("a", "b"))

# convert to column names tp upper case
change_colnames(data, toupper)
}
\seealso{
\code{\link{tag_colnames}}
}
