% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdif.R
\name{cvgdif}
\alias{cvgdif}
\title{Calculates the 'convergence difference'}
\usage{
cvgdif(x1, x2)
}
\arguments{
\item{x1}{first number}

\item{x2}{second number}
}
\value{
the 'convergence difference' as described above
}
\description{
\code{cvgdif} calculates the difference  between two numbers
\code{x1} and \code{x2} according to \code{|x_1 - x_2| / max(|x_2|, 1)}.
This difference is equivalent to the convergence test employed in the
package \code{macromod}.
}

