% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag2_diff2.R
\name{diff_ts}
\alias{diff_ts}
\title{Lagged differences}
\usage{
diff_ts(x, lag = 1, differences = 1, keep_range = TRUE)
}
\arguments{
\item{x}{A univariate or multivariate timeseries}

\item{lag}{an integer indicating which lag to use}

\item{differences}{an integer indicating the order of the difference.}

\item{keep_range}{If \code{TRUE} (the default), then  the output
timeseries has the same period range as the input timeseries.
Then the result timeseries will have \code{lag + differences - 1}
\code{NA} values at the beginning (\code{lead_ts}) or the end (\code{lag_ts}).
If \code{FALSE} then the result timeseries is \code{lag + differences - 1}
periods shorted.}
}
\description{
Returns suitaly lagged and iterated differences.
This function works similar as \code{\link[base]{diff}},
except that the function has an additional argument \code{keep_range}.
}
\examples{
x <- regts(1:10, start = "2018q3")
diff_ts(x)
diff_ts(x, lag = 2, keep_range = FALSE)
}
