% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disagg.R
\name{disagg}
\alias{disagg}
\title{Temporal disaggregation of time series using spline interpolation.}
\usage{
disagg(x, nfrequency, method = c("fmm", "periodic", "natural", "monoH.FC",
  "hyman", "nakn"), constraint = c("sum", "average", "last", "first"))
}
\arguments{
\item{x}{a \code{\link[stats]{ts}} object}

\item{nfrequency}{the frequency of the result. This should be higher than
the frequency of timeseries \code{x}.}

\item{method}{specifies the type of spline to be used.
Possible values are \code{"fmm"}, \code{"natural"}, \code{"periodic"},
\code{"monoH.FC"}, \code{"hyman"} and \code{"nakn"}. Can be abbreviated.}

\item{constraint}{Constraint on the high frequency result.
Possible values are \code{"sum"}, \code{"average"}, \code{"last"},
\code{"first"}.
Either the sum, the average, the first or the last value of the resulting
high-frequency series should be equal to the corresponding
low-frequency value.}
}
\description{
This function converts an timeseries to a timeseries with higher frequency,
for example a yearly timeseries to a quarterly timeseries. Missing values
are obtained by spline interpolation.
}
\details{
For each individual timeseries, trailing and leading \code{NA} values are
removed before interpolation. If the timeseries contains intermediate
\code{NA} values,  the resulting timeseries is set to \code{NA}.

Argument \code{method} can be used to select a spline interpolation
methods. All methods except \code{"nakn"} employ
the function code{\link[stats:splinefun]{spline}}
function of the \code{stats} package. More information
about the varous methods can be found in the documentation of that
package.

For method \code{"nakn"} the function {\link[cspline]{cspline}}
of package \code{cspline} is used. This method used a not-a-knot
boundary conditions, which implies that the first derivative
at the second and one but last point are continuous.
}
