% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_columns}
\alias{filter_columns}
\title{Select columns using  a filter}
\usage{
filter_columns(x, regex)
}
\arguments{
\item{x}{any R object that supports column selection (\code{data.frame},
\code{matrix}, \code{ts}, \code{regts})}

\item{regex}{a regular expression used to select column}
}
\value{
the selection of object \code{x}
}
\description{
This function selects columns of R objects that support column
selection (e.g. a \code{data.frame}, \code{matrix}, \code{ts} or
\code{regts})  using a
regular expression.
}

