% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod.R
\name{frequency.regperiod}
\alias{frequency.regperiod}
\title{Returns the frequency of a \link{regperiod} object}
\usage{
\method{frequency}{regperiod}(x, ...)
}
\arguments{
\item{x}{a \code{regperiod}}

\item{...}{additional arguments for future methods}
}
\value{
the frequency of the \code{regperiod}
}
\description{
Returns the frequency of a \link{regperiod} object
}
\examples{
p <- regperiod("2016Q1")
freq <- frequency(p)
}

