% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_periods.R
\name{get_periods}
\alias{get_periods}
\title{Return all periods in a  \code{period_range} or timeseries.}
\usage{
get_periods(x, ...)
}
\arguments{
\item{x}{a \code{\link{period_range}} or times series
(\code{\link[stats]{ts}} or \code{\link{regts}}).}

\item{...}{arguments passed to methods (currently not used).}
}
\description{
The periods in a \code{\link{period_range}} or timeseries are returned
as a \code{period} vector
}
\examples{
# example for period range
range <- period_range("2018m1/2018m3")
get_periods(range)

# example for timeseries
x <- regts(1:3, start = "2010Q4")
get_periods(x)
}
\seealso{
\code{\link{get_period_range}}
}
