% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_regperiod_range}
\alias{get_regperiod_range}
\title{Returns the period range of the time series as a \link{regperiod_range}
object.}
\usage{
get_regperiod_range(tsSEXP)
}
\arguments{
\item{x}{a \code{regts} or \code{ts}}
}
\value{
a \code{regperiod_range}
}
\description{
Returns the period range of the time series as a \link{regperiod_range}
object.
}

