% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_regperiod_range}
\alias{get_regperiod_range}
\title{Returns the \link{regperiod_range} of a timeseries.}
\usage{
get_regperiod_range(x)
}
\arguments{
\item{x}{a timeseries (\link{ts} or \link{regts})}
}
\value{
a \code{regperiod_range}
}
\description{
Returns the \link{regperiod_range} of a timeseries.
}
