% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod.R
\name{get_subperiod}
\alias{get_subperiod}
\title{Returns the subperiod of a \code{\link{regperiod}}}
\usage{
get_subperiod(x)
}
\arguments{
\item{x}{a \code{regperiod}}
}
\value{
the subperiod of a regperiod
}
\description{
This function returns the subperiod within a year.
For example, for \code{regperiod} \code{2011Q3} the function
returns 3.
}
\examples{
get_subperiod(regperiod("2010Q3"))
}
\seealso{
\code{\link{get_year}}
}

