% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod.R
\name{get_year}
\alias{get_year}
\title{Returns the year of a \code{\link{regperiod}}}
\usage{
get_year(x)
}
\arguments{
\item{x}{a \code{regperiod}}
}
\value{
the year
}
\description{
Returns the year of a \code{\link{regperiod}}
}
\examples{
get_year(regperiod("2010Q3"))
}
\seealso{
\code{\link{get_subperiod}}
}

