% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth}
\alias{growth}
\title{Return the relative growth rate of a timeseries}
\usage{
growth(x, n = 1, keep_range = TRUE)
}
\arguments{
\item{x}{a \code{\link[stats]{ts}} or \code{\link{regts}} object}

\item{n}{an integer indicating the period of relative change}

\item{keep_range}{if \code{TRUE} (the default), then the output
timeseries has the same period range as the input timeseries.
Then the result timeseries will have \code{n} \code{NA} values at the start.
If \code{FALSE} then the result timeseries is \code{n} periods
shorter than the input timeseries.}
}
\value{
a \code{regts} object with relative changes
}
\description{
Function \code{growth} computes the relative growth rate of a (multivariate)
timeseries. The one period relative growth rate  of a timeseries is defined
as \code{growth(x) = (x[t] - x[t-1]) / x[t-1]}.
The \code{n} period relative change of a timeseries is defined as
\code{growth(x,n) = (x[t] - x[t-n]) / x[t-n]}.

Note that \code{growth} divides the change by \code{ x[t-n]} and not by the
absolute value of \code{x[t-1]}.
This implies that the growth rate is positive when a negative timeseries
becomes more negative.
}
\examples{
x <- regts(rnorm(10), start = "2018Q1")
growth(x, keep_range = FALSE)
growth(x, 4)
}
\seealso{
\code{\link{rel2index}}
}
