% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_ts.R
\name{growth}
\alias{growth}
\title{Return the relative change of a timeseries}
\usage{
growth(x, n = 1, keep_range = TRUE)
}
\arguments{
\item{x}{a \code{\link[stats]{ts}} or \code{\link{regts}} object}

\item{n}{an integer indicating which lag to use}

\item{keep_range}{If \code{TRUE} (the default), then  the output
timeseries has the same period range as the input timeseries.
Then the result timeseries will have \code{lag} NA values at the start.
If \code{FALSE} then the result timeseries is \code{lag} periods
shorter than the input timeseries.}
}
\value{
a \code{regts} object with relative changes
}
\description{
Function \code{growth} computes the relative change of a timeseries.
The one period relative change of a timeseries is defined as:
growth(x) <- (x[t] - x[t-1]) / |x[t-1]|

The \code{n} period relative change of a timeseries is defined as:
growth(x) <- (x[t] - x[t-\code{n}]) / |x[t-\code{n}]|

The formula implies that when the timeseries decreases, the result will be
negative regardless of the sign of x. The function also works for
multivariate timeseries.
}
\examples{
x <- regts(rnorm(10), start = "2018Q1")
growth(x, keep_range = FALSE)
growth(x, 4)
}
