% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_ts.R
\name{index_ts}
\alias{index_ts}
\title{Construct an index timeseries by scaling}
\usage{
index_ts(x, base = NULL, scale = 100)
}
\arguments{
\item{x}{a \code{\link[stats]{ts}} of \code{\link{regts}} object}

\item{base}{a \code{\link{period}} or a
\code{\link{period_range}} specifying the base period, or an object that can
be coerced to a \code{period} or \code{period_range}.
By default the base period is the first period of the input timeseries.}

\item{scale}{the (average) value of the index series at the base period
(by default 100). This should be a positive number.}
}
\description{
This function scales a timeseries by dividing all observations by
one selected observation or by the mean of a range of observations.
The index series \code{i} is calculated from the input series \code{x} as

\if{html}{\out{<div class="sourceCode">}}\preformatted{i[t] = scale * x[t] / mean(x[base]),
}\if{html}{\out{</div>}}

where \code{scale} is usually 100 and \code{base} the base period, which can be
a single \code{period} or a \code{period_range} (by default the base period is the
first period of \code{x}).
If  \code{mean(x[base])} is negative, a warning is given and the (mean) value
of the resulting index series at the  base period will be \code{-scale}.
}
\examples{
\dontshow{
set.seed(123)
}
ab <- regts(matrix(1:18 + rnorm(18), ncol = 2), start = "2016Q1",
            names = c("a", "b"))

index_ts(ab)

index_ts(ab, base = "2017", scale = 1)

}
\seealso{
\code{\link{rel2index}} and \code{\link{pct2index}}
}
