% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod.R
\name{is.regperiod}
\alias{is.regperiod}
\title{Test if an object is a regperiod.}
\usage{
is.regperiod(x)
}
\arguments{
\item{x}{any R object}
}
\value{
\code{TRUE} if the object is a \code{\link{regperiod}}
}
\description{
Test if an object is a regperiod.
}
\examples{
p <- regperiod("2016Q1")
is.regperiod(p)
is.regperiod("2016Q1")
}

