% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod_range.R
\name{is.regperiod_range}
\alias{is.regperiod_range}
\title{Test if an object is a regperiod_range.}
\usage{
is.regperiod_range(x)
}
\arguments{
\item{x}{any R object}
}
\value{
\code{TRUE} if the object is a \code{regperiod_range}
}
\description{
Test if an object is a regperiod_range.
}

