% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_ts.R
\name{join_ts}
\alias{join_ts}
\title{Join two or more timeseries}
\usage{
join_ts(..., method = "union", suffixes)
}
\arguments{
\item{...}{two or more univariate or multivariate time series,
or objects which can coerced to time series}

\item{method}{method used to determine the time axis of the result.
Specify \code{"union"} for a total time coverage
(missing values are filled with \code{NA}) and \code{"intersect"}
to restrict the time axis to the time covered by all timeseries.}

\item{suffixes}{Suffixes appended to the column names for overlapping
columns. This argument is mandatory if the timeseries have overlapping
columns.}
}
\description{
Bind time series which have a common frequency.
}

