% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_ts.R
\name{join_ts}
\alias{join_ts}
\title{Join two or more timeseries}
\usage{
join_ts(..., union = TRUE, suffixes)
}
\arguments{
\item{...}{two or more univariate or multivariate time series,
or objects which can coerced to time series}

\item{union}{if \code{TRUE}, then the period range of the result
is the union of the period ranges of the joined objects
(missing values are filled with \code{NA}). If \code{FALSE},
then the period range of the result is the intersection of the period ranges
of the joined objects.}

\item{suffixes}{Suffixes appended to the column names for overlapping
columns. This argument is obligatory if the timeseries have overlapping
column names}
}
\description{
Join time series objects with a common frequency.
}
\seealso{
\code{\link{as.list.regts}}
}

