% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_ts_diff_ts.R
\name{lead_ts}
\alias{lead_ts}
\alias{lead_ts.ts}
\alias{lead_ts.default}
\title{Lead a Timeseries}
\usage{
lead_ts(x, n = 1, keep_range = TRUE, ...)

\method{lead_ts}{ts}(x, n = 1, keep_range = TRUE, ...)

\method{lead_ts}{default}(x, n = 1, keep_range = TRUE, ...)
}
\arguments{
\item{x}{a univariate or multivariate timeseries.
Can also be a vector, matrix or data frame (see details).}

\item{n}{the number of leads (in units of observations).
Must be a positive number.}

\item{keep_range}{if \code{TRUE} (the default), then the output
timeseries has the same period range as the input timeseries.
The result timeseries will have \code{n} \code{NA} values at the
beginning end.
If \code{FALSE} the period range of the result timeseries is shifted
by \code{n} periods. The result timeseries
starts and ends \code{n} periods earlier.}

\item{...}{further arguments to be passed to or from methods
(currently not used in package \code{regts})}
}
\description{
Compute the lead of a timeseries, shifting the observations
backwards by a given number of periods.
}
\details{
Vector, matrix and data frame arguments are first converted to a \code{regts}
with function \code{\link{regts}}. This conversion results in a timeseries
with frequency 1 and starting at year 1.

Function \code{lead_ts} is an alternative for function \code{\link[stats]{lag}}
in the \code{stats} package which computes both lags and leads.
By default in \code{lead_ts} the resulting timeseries has the same
period range as the input timeseries. In function \code{lag} the time base
is always shifted. \code{lead_ts(x, 1, keep_range = FALSE)} is the same as
\code{lag(x, 1)}
}
\examples{
x <- regts(1:10, start = "2018q3")
lead_ts(x)
lead_ts(x, k = 2, keep_range = FALSE)
}
\seealso{
\code{\link{lag_ts}} and \code{\link{diff_ts}}
}
