% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_df_to_regts.R
\name{long_df_to_regts}
\alias{long_df_to_regts}
\title{Converts a long data frame to a regts object.}
\usage{
long_df_to_regts(df, name_col = "name", period_col = "period",
  value_col = "value", label_col = "label", numeric = TRUE)
}
\arguments{
\item{df}{A long data frame (or \code{\link[tibble]{tibble}}) with at least
three columns with names specified with arguments \code{name_col}, \code{period_col}
and \code{value_col}.}

\item{name_col}{The name of the column with variable names (by default
\code{"name"}).}

\item{period_col}{The name of the column with periods. This columns
should contain data of a type that can be converted to \code{period} objects
by function \code{\link{period}}. The default is \code{"period"}.}

\item{value_col}{The name of the column with values (default \code{"value"}).}

\item{label_col}{The name of the column with labels (default \verb{"label}").
If not specified and a column named \code{"label"} exists, the texts in this
column are used to create timeseries labels. Specify \code{NULL} if you do not}

\item{numeric}{A logical. If \code{TRUE} (the default), the data in the column
with values are converted to numeric data.}
}
\value{
A \code{regts} object
}
\description{
The data frame should have at least three columns: one column with
the names of the timeseries, one column with periods and a third
column with the values. The names of these columns can be specified with
argument \code{name_col}, \code{period_col} and \code{value_col}, respectively
(the default column names are \code{"name"}, \code{"period"} and \code{"value"}).
\cr
Argument \code{label_col} can be used to specify the name of a column with
labels (default \code{"label"}). If not specified, we function checks whether
a column \code{"label"} exists. If this column exists, the labels are based on
the data in this column. If the column does not exist, the label column
is ignored. If the data frame does contain a column \code{"label"} but you do not
want to create labels specify \code{label_col = NULL}.
}
\examples{
df <- tibble::tribble(
~name,  ~period,  ~value,  ~description,
"a",    "2015Q3", 1.2 ,     "Var a",
"a",    "2016Q1", 1.5 ,     "Var a",
"b",    "2015Q3", 15 ,      "Var b",
"b",    "2015Q4", 20,       "Var b"
 )
long_df_to_regts(df, label_col = "description")
}
