% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_df_to_regts.R
\name{long_df_to_regts}
\alias{long_df_to_regts}
\title{Converts a long data frame to a regts object.}
\usage{
long_df_to_regts(df, name_col = "name", period_col = "period",
  value_col = "value", label_col = "label", numeric = TRUE)
}
\arguments{
\item{df}{A long data frame (or \code{\link[tibble]{tibble}}) with at
least three columns containing names, periods and values. The names
of these columns can be specified with arguments \code{name_col}, \code{period_col}
and \code{value_col}.}

\item{name_col}{The name of the column with variable names
(by default \code{"name"}).}

\item{period_col}{The name of the column with periods
(by default \code{"period"}). This column
should contain data of a type that can be converted to \code{period}
objects by the function \code{\link{period}}.}

\item{value_col}{The name of the column with values (default \code{"value"}).}

\item{label_col}{The name of the column with labels (default \code{"label"}).
If the column exists, labels are extracted from it. If the column does
not exist and \code{label_col} was explicitly set, an error is thrown.
If the column does not exist and the default was used, the result
is returned without labels. Specify \code{NULL} if a column \code{"label"}
exists but you want to ignore the labels.}

\item{numeric}{A logical. If \code{TRUE} (the default), the data in the column
with values are converted to numeric data.}
}
\value{
A \code{\link{regts}} object.
}
\description{
The data frame should have at least three columns: a column with
the names of the timeseries, a column with periods and another
column with the values. The names of these columns can be specified with
arguments \code{name_col}, \code{period_col} and \code{value_col}, respectively
(the default column names are \code{"name"}, \code{"period"} and \code{"value"}).
\cr \cr
The data frame can optionally contain an additional column with labels. The
name of this column can be specified with argument \code{label_col}. If this
column exists, the resulting \code{regts} object is equipped with the labels from
this column.
}
\examples{
df <- tibble::tribble(
~name,  ~period,  ~value,  ~description,
"a",    "2015Q3", 1.2 ,     "Var a",
"a",    "2016Q1", 1.5 ,     "Var a",
"b",    "2015Q3", 15 ,      "Var b",
"b",    "2015Q4", 20,       "Var b"
 )
long_df_to_regts(df, label_col = "description")
}
\seealso{
\code{\link{as_data_frame}}.
}
