% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movav.R
\name{movav}
\alias{movav}
\title{Moving average of a timeseries}
\usage{
movav(x, max_lag = 0L, max_lead = 0L, keep_range = TRUE)
}
\arguments{
\item{x}{a \code{\link[stats]{ts}} or \code{\link{regts}} object}

\item{max_lag}{the maximum lag}

\item{max_lead}{the maximal lead}

\item{keep_range}{If \code{TRUE} (the default), then  the output
timeseries has the same period range as the input timeseries.
The result will have \code{max_lag} \code{NA}s at the left side and
\code{max_lead} NAs at the right side). If \code{FALSE}, then the
result has a shorter period rang than the input timeseries
(it starts \code{max_lags} periods later and ends \code{max_lead} periods
earlier).}
}
\value{
a \code{regts} object with the moving average values
}
\description{
The moving average smoother calculates for each observation the average
of the observations in a range around that observation.
For example, the backwards moving average
of order 3 is given by  \eqn{A[t] = (x[t-2] + x[t-1] + x[t]) / 3}.
In general, the moving average is calculated as

\eqn{A[t] = (x[t - p] + x[t - p + 1] + \dots + x[t] + \dots +
x[t + q - 1] + x[t + q]) / n},

where \eqn{p} and \eqn{q} are the
maximum lag and lead, respectively, and \eqn{n = p + q + 1} is the order
of the moving average.
}
\examples{
x <- regts(rnorm(10), start = "2018Q1")

# backward moving average of order 3
movav(x, max_lag = 2)

# centered moving average of order 3
movav(x, max_lag = 1, max_lead = 1, keep_range = FALSE)
}
