% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movav.R
\name{movav}
\alias{movav}
\title{Moving average}
\usage{
movav(x, max_lag = 0L, max_lead = 0L, na_pad = TRUE)
}
\arguments{
\item{x}{a a \code{\link[stats]{ts}} of \code{\link{regts}} object}

\item{max_lag}{a numeric value specifying the maximum lag
for the moving averaging calculation. This should be a negative number or zero.}

\item{max_lead}{a numeric value the bound in the summation (see details).
This should be a positive number or zero.}

\item{na_pad}{If \code{TRUE} (the default), then  the output
timeseries has the same period range as the input timeseries.
NA values are padded to the left and the right of calculated average (
there wille be \code{-from} \code{NA}s at the left side and \code{to} NAs
to the right side)}
}
\value{
a \code{regts} object with the moving average values
}
\description{
Function \code{movav} calculates the moving average of a timeseries.
The moving average of a timeseries \eqn{x[t]} at time \eqn{t}
is defined as \eqn{A[t] = (x[t - p] + x[t - p + 1] + ... +
x[t + q - 1] + x[t + q]) / n}, where \eqn{-p} is the lower bound
and \eqn{q} the upper bound of the summation.
}
