% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_matrix.R
\name{numeric_matrix}
\alias{numeric_matrix}
\title{Convert a data frame to a numeric matrix.}
\usage{
numeric_matrix(x, dec = ".")
}
\arguments{
\item{x}{a data frame}

\item{dec}{decimal separator in number strings. The default is \code{"."}}
}
\description{
Returns the matrix obtained by converting all the variables in a data frame
to numeric mode and then binding them together as the columns of a matrix.
Factors are converted to characters.
}
\details{
This function is similar to the function \code{\link[base]{data.matrix}}
of the base package, except that:
\itemize{
\item it is possible to specify a decimal separator used to convert strings to
numeric values
\item factors are first converted to characters and then to numeric values
\item if the data frame contains strings that cannot be converted
to numerical values, then \code{numeric_matrix} gives a warning about
the first 10 character strings that could not be succesfully converted.
}
}
\examples{
df <- data.frame(a = c("1.123", "x", NA), b = c("1", "", "john"),
                 c = 10 : 12)
numeric_matrix(df)
}
