% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{period}
\alias{period}
\title{Create a \code{period} object based on a string or an integer.}
\usage{
period(x, frequency = NA)
}
\arguments{
\item{x}{a single string or integer}

\item{frequency}{frequency of the period object.
Argument \code{frequency} is mandatory if a general period format
such as "2011-1" has been specified}
}
\value{
a \code{period} object
}
\description{
Possible string formats are for example \code{"2010Q2"},
\code{"2010M2"}, \code{"2011"} or \code{"2011-1"}.
}
\examples{
period("2010Q3")
period("2010-4", frequency = 3)
period(2015)
}
