% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{period}
\alias{period}
\alias{as.period}
\title{Create a \code{\link{period}} object}
\usage{
period(x, frequency = NA)

as.period(x, ...)
}
\arguments{
\item{x}{a character, numeric,
\code{\link[base]{Date}}, \code{\link[base:DateTimeClasses]{POSIXct}}
or \code{\link[base:Dates]{POSIXlt}} vector.}

\item{frequency}{frequency of the period. Argument \code{frequency} is
mandatory if the frequency cannot be inferred from \code{x} (for example
\code{"2017-2"} could be a quarter, month, etc.)}

\item{...}{additional arguments to be passed to or from methods (currently
not used in package \code{regts})}
}
\value{
a \code{period} vector if all periods have the same frequency,
otherwise a list of \code{period} objects.
}
\description{
Function \code{period} creates a \code{period} object
based on a character string or numeric scalar.
Possible character string formats are for example \code{"2017Q2"},
\code{"2017m2"}, \code{"2017"}, \code{"2017-2"},
\code{"aug 2017"} or \code{"august-2017"}.
Possible numeric formats are for example \code{2017} or
\code{2017.25} (the second quarter or the fourth month of 2017).
The function also accepts a \code{\link{Date}},
\code{\link[base:DateTimeClasses]{POSIXct}} or
\code{\link[base:DateTimeClasses]{POSIXlt}} argument.
See Details.
\cr\cr
Function \code{as.period}
coerces an R object to a \code{period} object if possible.
}
\details{
The function \code{period} accepts a character string or
a numeric scalar as arguments. The specific format is described below.

\strong{string format}

The format for yearly periods is for example \code{"2017"}
or \code{"2017Y"} (the suffix \code{"Y"} is optional).

The standard format format for quarterly periods is
for example \code{"2017Q3"}. Alternative formats
such as \code{"2017 3Q"} and \code{"2017.3Q"}
are also recognized. The separator between the year and the
quarter can be a blank or a dot, as in the previous examples,
but also a forward slash (\code{"/"}) and underscore (\code{"_"})
are allowed.

The format for monthly periods is similar as that of
quarterly periods, except that the \code{"Q"} is replaced by
\code{"M"}. Monthly periods may also be specified
with a month name (possibly abbreviated) and year
(e.g. \code{"aug 2017"}, \code{"2018-August"}). The parser only understands
English month names.

Periods with other frequencies than year, quarter and month can be specified
as for example \code{"2017-2"}. Alternative
separators (blank, dot, etc.) are possible. In this case argument
\code{frequency} should be specified.

The string format is case insensitive, and may be prefixed
with \code{"Y"} or \code{"T"}. Thus for exampe \code{"t2017q3"}
is also an allowed period string.

\strong{numeric format}

An integer number, such as \code{2017} specifies a year,
or the first subperiod in a year if argument \code{frequency} has been
specified.

If the numeric has a non-zero fractional part,
then argument \code{frequency} is mandatory,
For example, the numeric \code{2017.25}
can specify the second quarter of 2017 or the fourth month of 2017.

\strong{\code{Date}, \code{POSIXct} and \code{POSIXlt}}

The function also accepts a \code{\link{Date}},
\code{\link[base:DateTimeClasses]{POSIXct}} or
\code{\link[base:DateTimeClasses]{POSIXlt}} argument.
By default the function converts this object to a \code{period} with
frequency month. It is possible to specify another output frequency,
provided that this frequency is a divisor of 12.
}
\examples{
period("2010Q3")
period("2010-2", frequency = 3)
period(2015)
period(2010.25, frequency = 4)

# examples for as.period
as.period("2010q3")
p <- period("2010m11")
as.period(p)

# example with a Date object
d <- Sys.Date()
period(d)

# create a sequence of period objects
seq(period("2018q2"), period("2019q4"))
}
\seealso{
\code{\link{period_range}} and \code{\link{seq}}
}
