% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_range.R
\name{period_range}
\alias{period_range}
\alias{as.period_range}
\title{Create a \code{\link{period_range}} object.}
\usage{
period_range(start = NULL, end = NULL, frequency = NA)

as.period_range(x, frequency = NA, ...)
}
\arguments{
\item{start}{the first period (a \code{period}, an object that can be coerced
to a \code{period}, or by default \code{NULL}). If \code{start} is \code{NULL}
the lower bound of the period range is undetermined. \code{start} can also
be a character string specifying a period range, for example
\code{"2010Q2/2011Q3"}).}

\item{end}{the last period (a \code{period}, an object that can be coerced
to a \code{period}, or by default \code{NULL}).
If \code{end} is \code{NULL}, the upper bound of the period range is undetermined.}

\item{frequency}{frequency of the period objects. This argument is mandatory
if argument \code{start} or \code{end} is a character with general period format
without frequency indicator (e.g. \code{"2011-1"})}

\item{x}{an R object}

\item{...}{additional arguments to be passed to or from methods (currently
not used in package \code{regts})}
}
\value{
a \code{period_range} object
}
\description{
A \code{period_range} object represents an interval of periods,
for example a period from \code{"2017Q2"} to \code{"2019Q3"}.
Function \code{period_range}  creates a \code{period_range} from
a single character string (e.g. \code{"2017Q2/2019Q3"}, see Details)
or from two R objects that can be coerced to \code{period} objects.
\cr\cr
Function \code{as.period_range} coerces an R object to a \code{period_range}
if possible.
}
\details{
It is possible to create a \code{period_range} from
a single string specifying a period range, for example
\code{"2017Q2/2019Q3"}. For this format, the first and last
period are separated by \code{"/"}. The first and last period
are specified according to the same format recognized by function
\code{\link{period}}. The first or last period may be omitted
(e.g. \code{"2017Q3/"})), in that case the period range
has no lower or upper bound. The string format is case insensitive.
}
\examples{
# two methods to create a period_range from 2010Q2 to 2016Q3
period_range("2010Q2", "2016Q3")
period_range("2010Q2/2016Q3")

# create a period_range for the first 5 quarters after 2013Q2
start <- period("2013q3")
period_range(start, start + 5)

# create a period_range up to 2010Q2 with no lower bound
period_range(end = "2010q2")

# create a period_range for a timeseries with frequency 2 (half year)
period_range("2010-2", "2016-2", frequency = 2)

# convert a period object to a period_range with equal start and end period
p <- period("2010Q2")
as.period_range(p, p)

# create a month range starting at the month 1000 days before
# the current day and ending at the current month.
today <- Sys.Date()
period_range(today - 1000, today)
}
\seealso{
\code{\link{period}}, \code{\link{nperiod}},
\code{\link{start_period}}, \code{\link{end_period}}, \code{\link{seq}} and
\code{\link{get_periods}}.
}
