% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regperiod_range_utils.R
\name{range_intersect}
\alias{range_intersect}
\title{Calculates the intersection of two period_range objects.}
\usage{
range_intersect(range1, range2)
}
\arguments{
\item{range1}{a \code{period_range} object}

\item{range2}{another \code{period_range} object}
}
\value{
the intersection of \code{range1} and \code{range2}.
It returns NULL if there is no intersection
}
\description{
This function calculates the intersection of two period_range
objects. The start and end periods of the period_range objects may
not be NULL.
}
\examples{
range1 <- period_range("2016Q1", "2017Q4")
range2 <- period_range("2017Q1", "2018Q2")
range_intersect <- range_intersect(range1, range2)
}
\seealso{
\code{\link{union}}
}
