% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regts.R
\name{regts}
\alias{regts}
\title{Create a \code{regts} timeseries object}
\usage{
regts(data, start, end = NULL, frequency = NA, names = NULL,
  labels = NULL)
}
\arguments{
\item{data}{a vector or matrix of the observed time-series values. A data frame will be
coerced to a numeric matrix via \link{data.matrix}. (See also the description of the
function \link{ts} of the \link{stats} package).}

\item{start}{the starting period as a  \link{regperiod} object or a character string that can
be converted to a \code{regperiod} object}

\item{end}{the end period as a  \link{regperiod} object or a character string that can
be converted to a \code{regperiod} object. If not specified, then the end period is calculated
from the dimension of \code{data}}

\item{frequency}{the frequency of the timeseries. This argument should only be specified if
the start or end period is specified with a general period format without period indicator,
e.g. \code{"2011-3"}}

\item{names}{a character vector of names for the series: defaults to the
colnames of data, or \code{"Series 1"}, \code{"Series 2"} ..., if the data does
not have colnames. In contrast to the function \link{ts},  \code{names}
is also used if the result is a single timeseries.}

\item{labels}{a character vector of labels (descriptions of the timeseries)}
}
\value{
a \code{regts} object
}
\description{
The \code{regts} class is an extension of the \link{ts} class of the \link{stats} package.
Working with \code{regts} makes it easier to select periods. Another difference is
related to the representation of univariate timeseries.  A \code{regts} object
always has \link{colnames}, also for univariate timeseries. A univariate \code{ts} object
only has column names when the input data was a one-dimensional matrix.
}
\examples{
# univariate timeseries
ts1 <- regts(1:10, start = "2010Q4")

# multivariate timeseries
ts2 <- regts(matrix(1:9, ncol = 3), start = "2010Q4", names = c("a", "b", "c"))

# multivariate timeseries with labels
ts3 <- regts(matrix(1:9, ncol = 3), start = "2010Q4", names = c("a", "b", "c"),
             labels = paste("Timeseries", c("a", "b", "c")))

# create a half-yearly timeseries, because ‘end’ is specified the length
# of the timeseries is smaller than the length of data (10).
ts4 <- regts(1:10, start = "2010-1", end = '2011-2', frequency = 2)
}
\seealso{
The function \link{is.regts} can be used to test if an object is a \code{regts}.

The S3 generic \link{as.regts} can be used to coerce an R object to a \code{regts}. There are currently
methods for \link{ts} and \link{data.frame}.

\link{as.data.frame.regts} and \link{as.list.regts} can be used
to convert \code{regts} to a \link{data.frame} or a \link{list}.

Information about the time period of the timeseries can be obtained
with the functions \link{get_regperiod_range}, \link{start_period}
and \link{end_period}.

See also the description of the functions for handling labels
(\link{ts_labels} and \link{update_ts_labels}).
}

