% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regts.R
\name{regts}
\alias{regts}
\title{Create a \code{regts} timeseries object}
\usage{
regts(data, start, end, period, frequency = NA, names = colnames(data),
  labels = NULL)
}
\arguments{
\item{data}{a vector or matrix of the observed time-series values. A data frame
will be coerced to a numeric matrix via \code{\link{data.matrix}}.
(See also the description of the
function \code{\link[stats]{ts}} of the \code{\link{stats}} package).}

\item{start}{the starting period as a  \code{\link{regperiod}} object or a
character string that can be converted to a \code{regperiod} object.
If not specified, then the start period is calculated
from argument \code{end} and the dimension of \code{data}}

\item{end}{the end period as a  \link{regperiod} object or a character string
that can be converted to a \code{regperiod} object. If not specified, then
the end period is calculated from argument \code{start} and
the dimension of \code{data}}

\item{period}{the period range as a \code{\link{regperiod_range}} object or a
character string that can be converted to a \code{regperiod_range} object.
This argument replaces arguments \code{start} and \code{end}.}

\item{frequency}{the frequency of the timeseries. This argument should only be
specified if the start, end or period argument is specified with a general
period format without period indicator, e.g. \code{"2011-3"}}

\item{names}{a character vector with the column names for the series
if \code{data} is a matrix or data frame. Defaults to the colnames of data.}

\item{labels}{a character vector of labels (descriptions of the timeseries)}
}
\value{
a \code{regts} object
}
\description{
The \code{regts} class is an extension of the \code{\link[stats]{ts}} class
of the \link{stats} package. Working with \code{regts} makes it easier to
select periods.
}
\examples{
# univariate timeseries
ts1 <- regts(1:10, start = "2010Q4")


# multivariate timeseries
ts2 <- regts(matrix(1:9, ncol = 3), start = "2010Q4", names = c("a", "b", "c"))

# multivariate timeseries with labels
ts3 <- regts(matrix(1:9, ncol = 3), start = "2010Q4", names = c("a", "b", "c"),
             labels = paste("Timeseries", c("a", "b", "c")))

# multivariate timeseries with period
range <- regperiod_range("2016Q1", "2017Q4")
ts4 <- regts(matrix(1:16, ncol = 2), period = range, names = c("a", "b"))

# create a half-yearly timeseries; because argument end is specified the
# length of the timeseries is smaller than the length of data (10).
ts5 <- regts(1:10, start = "2010-1", end = '2011-2', frequency = 2)
}
\seealso{
The function \code{\link{is.regts}} can be used to test if an object is a
\code{regts}.

The S3 generic \code{\link{as.regts}} can be used to coerce an R object to a
 \code{regts}. There are currently methods for \code{\link[stats]{ts}} and
 \code{\link{data.frame}}.

\code{\link{as.data.frame.regts}} and \code{\link{as.list.regts}} can be used
to convert \code{regts} to a \code{\link{data.frame}} or a \code{\link{list}}.

Function \code{\link{cbind}} can be used to bind two or more
timeseries objects and create a multivariate \code{regts}.

Information about the time period of the timeseries can be obtained
with the functions \code{\link{regperiod_range}},
\code{\link{start_period}} and \code{\link{end_period}}.

See also the description of the functions for handling labels
(\code{\link{ts_labels}} and \code{\link{update_ts_labels}}).
}

