% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{rel2index/pct2index}
\alias{rel2index/pct2index}
\alias{rel2index}
\alias{pct2index}
\title{Calculate an index timeseries from a timeseries with relative or
percentage changes.}
\usage{
rel2index(x, base_period = start_period(x) - 1, scale = 100)

pct2index(x, base_period = start_period(x) - 1, scale = 100)
}
\arguments{
\item{x}{a \code{\link[stats]{ts}} or \code{\link{regts}}}

\item{base_period}{base period of the index timeseries (a \code{\link{period}}
object or an object that can be coerced to a \code{period} object)}

\item{scale}{the value of the index series at the base period}
}
\description{
The relative change of a timeseries X[t]
is defined as \code{x[t] = (X[t] - X[t-1]) / |X[t-1]|}.
Suppose that \code{x[t]} is given but \code{X[t]} is not known.
Then it is possible to calculate the index series
\code{i[t] = s * X[t] / X[t*]}, where \code{s} is arbitrary scale and
\code{t*} an arbitrary base period.
Function \code{rel2index} computes this index series \code{i[t]}
\cr\cr
Similarly, function \code{pct2index} computes the index series
for a timeseries of percentage changes, defined as
\code{x[t] = 100 (X[t] - X[t-1]) / | X[t-1]|}.
}
\section{Functions}{
\itemize{
\item \code{rel2index}: Calculates an index timeseries from a
timeseries with relative changes

\item \code{pct2index}: Calculates an index timeseries from a
timeseries with percentage changes
}}

\examples{
ts1 <- regts(abs(rnorm(10)), start = "2010Q2")
print(rel2index(ts1))
print(rel2index(ts1, base_period = "2010Q3", scale = 1))
}
\seealso{
\code{\link{index_ts}}
}
