% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel2index.R
\name{rel2index/pct2index}
\alias{rel2index/pct2index}
\alias{rel2index}
\alias{pct2index}
\title{Calculate an index timeseries from a timeseries with relative or
percentage changes.}
\usage{
rel2index(x, base = start_period(x) - 1, scale = 100, keep_range = TRUE)

pct2index(x, base = start_period(x) - 1, scale = 100, keep_range = TRUE)
}
\arguments{
\item{x}{a \code{\link[stats]{ts}} or \code{\link{regts}} (can also be a multivariate
timeseries)}

\item{base}{base period of the index timeseries (a \code{\link{period}}
object or an object that can be coerced to a \code{period} object)}

\item{scale}{the value of the index series at the base period (by default 100)}

\item{keep_range}{if \code{TRUE} (the default), then the output
timeseries has the same period range as the input timeseries.
If \code{FALSE} then the result timeseries starts 1 period earlier.}
}
\description{
This is the inverse of function \code{\link{growth}}.
The relative change, or growth, of a timeseries x\link{t}
is defined as \code{growth[t] = (x[t] - x[t-1]) / |x[t-1]|}.
Suppose that \code{growth[t]} is given but \code{x[t]} is not known.
Then it is possible to calculate the index series
\code{i[t] = s * x[t] / x[t*]}, where \code{s} is arbitrary scale and
\code{t*} an arbitrary base period.
Function \code{rel2index} computes this index series \code{i[t]}
\cr\cr
Similarly, function \code{pct2index} computes the index series
for a timeseries of percentage changes, defined as
\code{100 (x[t] - x[t-1]) / | x[t-1]|}.
}
\section{Functions}{
\itemize{
\item \code{rel2index}: Calculates an index timeseries from a
timeseries with relative changes

\item \code{pct2index}: Calculates an index timeseries from a
timeseries with percentage changes
}}

\examples{
ts1 <- regts(abs(rnorm(10)), start = "2010Q2")
print(rel2index(ts1))
print(rel2index(ts1, base = "2010Q3", scale = 1, keep_range = TRUE))
}
\seealso{
\code{\link{index_ts}} and \code{\link{growth}}
}
