% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{rel2index}
\alias{rel2index}
\alias{pct2index}
\title{Calculates an index timeseries from a timeseries with relative or
percentage changes}
\usage{
rel2index(x, base_period = start_period(x) - 1, scale = 100)

pct2index(x, base_period = start_period(x) - 1, scale = 100)
}
\arguments{
\item{x}{a \code{\link[stats]{ts}} or \code{\link{regts}}}

\item{base_period}{base period of the index timeseries (a \code{\link{period}}
object or an object that can be coerced to a \code{period} object)}

\item{scale}{the value of the index series at the base period}
}
\description{
A timeseries of relative changes \eqn{x_t} is relatived to the original
timeseries \eqn{X_t} by the relation \eqn{x_t = (X_t - X_{t-1}) / | X_{t-1}|}.
Function \code{rel2index} computes the index series
\eqn{\tilde{X}_t = s X_t / X_{t^*}}, where \eqn{s} is the scale and
\eqn{t^*} is the base period
}
\section{Functions}{
\itemize{
\item \code{rel2index}: Calculates an index timeseries from a timeseries with relative changes

\item \code{pct2index}: Calculates an index timeseries from a timeseries with
percentage changes
}}

\examples{
ts1 <- regts(abs(rnorm(10)), start = "2010Q2")
print(rel2index(ts1))
print(rel2index(ts1, base_period = "2010Q3", scale = 1))

}
