% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_range.R
\name{select_range}
\alias{select_range}
\title{Select a period range from a timeseries}
\usage{
select_range(x, range)
}
\arguments{
\item{x}{A timeseries object (a \code{regts} or a \code{ts}).}

\item{range}{The period range to select, as a \code{\link{period_range}}
object or a character string that can be converted to a \code{period_range}.
Partial ranges like \code{"2011/"} (from 2011 onwards) are supported.}
}
\value{
A \code{regts} object containing the selected period range.
}
\description{
This function selects a part of a \code{regts} or \code{\link[stats]{ts}}
object based on a \code{\link{period_range}}.
This function is convenient for use in a pipe.
}
\examples{
data <- regts(1:10, start = "2010Q1")
select_range(data, "2010Q2/2010Q3")

# Select data from the beginning of the year 2011 using a pipe.
data |> select_range("2011/")
}
\seealso{
\link{select} and \link{select_columns}.
}
