% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_df.R
\name{transpose_df}
\alias{transpose_df}
\title{Transpose a \code{data.frame}}
\usage{
transpose_df(x, colname_column, label_column)
}
\arguments{
\item{x}{a data frame}

\item{colname_column}{the name or the index of the column that contains
the column names of the transposed data frame. By default the
row names of the original data frame are used as column names of the new data
 frame.}

\item{label_column}{a numeric or character vector with the
indices or the names of the columns that contains
the row labels. If this is a vector of
length larger than 1, then the texts in the columns are combined
to create single labels for the columns of the transposed data frame.
By default labels are ignored.}
}
\value{
the transposed data frame
}
\description{
The function transposes a \code{\link[base]{data.frame}}.
If the input data frame has column labels (i. e., column vectors with an
attribute \code{"label")}), then the first column of the returned data frame
will contain the labels.
With argument \code{label_column} you can specify a column that will be
used to create column labels for the output data frame.
These labels are visible in de data viewer.
}
\examples{
df <- data.frame(variables = c("a", "b"),
                 labels = c("Variabele a", "Variable b"),
                 x = 1:2, y = 10:11)
df_t <- transpose_df(df, colname_column = 1, label_column = 2)
print(df_t)
print(transpose_df(df_t))
}
