% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdif.R
\name{tsdif}
\alias{tsdif}
\title{Calculate the differences between two multivariate timeseries
objects}
\usage{
tsdif(x1, x2, tol = 0, fun = function(x1, x2) (x1 - x2))
}
\arguments{
\item{x1}{the first timeseries (a  \code{\link{regts}} or
\code{\link[stats]{ts}} object).}

\item{x2}{the second timeseries (a \code{regts} or \code{ts} object).}

\item{tol}{difference tolerance (by default zero). Differences with absolute
values smaller than or equal to \code{tol} are ignored.}

\item{fun}{function to compute differences. This function should accept
two arguments (two numbers) for which the difference is computed.
By default the normal difference (\eqn{x_1 - x_2}) is computed. A useful
function for computing differences is \code{\link{cvgdif}}, which computes
relative differences if the absolute value of \code{x2} is larger than 1.}
}
\value{
a list with class "tsdif", with the following components
\item{equal}{\code{TRUE} if \code{x1} and \code{x2} have the same column
names and period ranges,
and if all differences are smaller than or equal to \code{tol}}
\item{difnames}{The names of the timeseries with differences
larger than \code{tol}}
\item{dif}{A \code{regts} with the computed differences,
or \code{NULL} if there are no differences larger than \code{tol}.
Only timeseries with differences larger than \code{tol} are included.
Leading and trailing rows with differences less than \code{tol} have also
been removed.}
\item{dif_table}{A \code{data.frame} with all differences greater
than the tolerance. The data frame contains four columns: \code{name}, \code{period},
\code{value1}, \code{value2} and \code{dif}.}
\item{maxdif}{A \code{data.frame} with the maximum differences. For each
timeseries the maximum difference is determined. Column \code{maxdif}
contains the maximum differences and
column \code{period} the periods at which these maximum difference occur.
Columns \code{value1} and \code{value2} are the values in timeseries \code{x1} and \code{x2.}
The rows of the data frame are ordered with decreasing order of \code{maxdif},
so the timeseries with the largest maximum differences come first. \code{NA}
values come first.}
\item{common_names}{the names of the common columns}
\item{missing_names1}{The names of columns present in \code{x2} but missing
in \code{x1}}
\item{missing_names2}{The names of columns present in \code{x1} but missing
in \code{x2}}
\item{period_range1}{The period ranges of \code{x1} as a
\code{\link{period_range}} object}
\item{period_range2}{The period ranges of \code{x2} as a
\code{period_range} object}
\item{common_range}{The intersection of the period ranges }
\item{ranges_equal}{A logical indicating whether the period ranges of
\code{x1} and \code{x2} differ}
\item{ts_names}{a character string giving the names of the two input
timeseries}
\item{tol}{The tolerance parameter}
\item{fun}{a character string specifying the supplied function \code{fun},
or \code{NULL} if \code{fun} has not been specified}
}
\description{
This function can be used to compare two multivariate timeseries objects.
The result is a list with a \code{\link{regts}} component with the computed
differences or \code{NULL} if there are no differences.
The function returns a list with the differences, the names of columns with
differences larger than a specified tolerance, and the names of the columns
present in one object but missing in the other object. The return value also
includes differences in the period ranges.
}
\details{
This function calculates the difference between common columns
of two timeseries objects \code{x1} and \code{x2}.
The two timeseries must have the same frequency, but may have a different
period range. The difference is computed for the intersection of the two
period ranges. Two \code{NA} or two \code{NaN} values are considered to be
equal. A \code{NA} value is not equal to a \code{NaN} value.
The function also returns missing column names in one of the two objects.

The return value of the function is an object of class \code{tsdif}. When this
object is printed, a short summary of the result of the
comparison is presented: the names of timeseries with differences,
the names of timeseries present
in one timeseries object but missing in the other object, and a table of the
maximum differences. For the table of maximum differences, the maximum
difference is determined for each timeseries separately,
and the maximum differences are printed in decreasing order, together with
the periods for which the maximum difference occurs for the specific
timeseries.

The print result is controlled by two options: \code{regts_max_difnames} and
\code{regts_max_maxdif}.  The first option, \code{regts_max_difnames} (default 50)
determines the maximum number of timeseries names printed (the names
of timeseries with differences and the names of timeseries missing in the
first or second timeseries object). Option \code{regts_max_maxdif} (default 10)
determines the maximum number of
maximum differences printed. The options can be modified with function
\code{\link[base]{options}} (e.g. \code{options(regts_max_maxdif = 20)} and
\code{options(regts_max_difnames = 1000)}. Function \code{\link[base]{getOption}}
can be used to check the current values of these options
(e.g. \code{getOption("regts_max_maxdif")}).
}
\examples{
# create two timeseries x1 and x2
x1 <- regts(matrix(data = rep(1:27), nc = 3), start = "2008Q4",
           names = c("a", "b", "c"))
x2 <- x1 + 0.001
colnames(x2) <- c("a", "b", "d")

# calculate and print the differences
dif1 <- tsdif(x1, x2)
print(dif1)

# use the function cvgdif (convergence difference)
dif2 <- tsdif(x1, x2, fun = cvgdif)

# calculate differences with tol = 1e-4, and print the
# names of the timeseries with differences larger than tol
dif3 <- tsdif(x1, x2, tol = 1e-4, fun = cvgdif)
print(dif3$difnames)

#
# example for timeseries objects with many columns
#

# create two timeseries objects with 100 timeseries
x1 <- regts(matrix(rnorm(10 * 100), ncol = 100), start =  "2018Q1",
            names = paste0("x", 1:100))
x2 <- x1

# Make x2 different from x1 at 20 random locations:
smpl <- sample.int(length(x1), 20)
x2[smpl] <- x2[smpl] + 1:length(smpl)

# set option regts_max_maxdif to ensure that all 20 differences are printed:
options(regts_max_maxdif = 20)
print(tsdif(x1, x2))
}
\seealso{
\code{\link{regts}}
}
