% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_regts.R
\name{tsupdate}
\alias{tsupdate}
\title{Updates a multivariate timeseries with another multivariate timeseries
object}
\usage{
tsupdate(x1, x2, method = c("replace", "tsupd", "tsupdna", "tsupdval"))
}
\arguments{
\item{x1}{the first timeseries (a multivariate \code{\link{regts}} or
\code{\link[stats]{ts}} object).}

\item{x2}{the second timeseries (a multivariate \code{regts} or \code{ts} object).}

\item{method}{four different ways to update the timeseries.
By default the timeseries are replaced. This behaviour can be changed by
using one of the other methods. See details.}
}
\value{
an updated multivariate regts object.
}
\description{
This function can be used to update a multivariate (reg)ts object with
another multivariate (reg)ts object.
The result is an updated multivariate regts object.
}
\details{
The timeseries can be updated in four different ways:

\code{replace} the first timeseries are replaced by the second timeseries

\code{tsupd} the first timeseries are updated with the second timeseries.
The two timeseries must have the same frequency, but may have a different
period range. The update is computed for the union of the two period ranges
(for the non-overlapping periods the result will be \code{NA}).

\code{tsupdna} if method \code{tsupdna} is selected instead of \code{tsupd},
only NA values in the first timeseries will be updated

\code{tsupdval} if method \code{tsupdval} is selected instead of \code{tsupd},
the values in the first timeseries are only replaced with valid (i.e. non-NA)
values from the second timeseries.

These methods are only employed at the common columns in both timeseries.
The non overlapping columns in both timeseries are added to the result.
}
\examples{
library(regts)

x1 <- regts(matrix(data = rep(1:9), nc = 3), period = "2000/2002",
            names = c("a", "b", "c"))
x2 <- regts(matrix(data = rep(10:15), nc = 3), period = "2000/2001",
            names = c("a", "c", "d"))
tsupdate(x1, x2, method = "tsupd")

}
\seealso{
\code{\link{regts}}
}
