% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ts.R
\name{update_ts}
\alias{update_ts}
\title{Update a timeseries with another timeseries object}
\usage{
update_ts(x1, x2, method = c("upd", "updna", "updval", "replace"))
}
\arguments{
\item{x1}{the first timeseries (a \code{\link{regts}} or
\code{\link[stats]{ts}} object).}

\item{x2}{the second timeseries (a \code{regts} or \code{ts} object).}

\item{method}{four different ways to update the timeseries.
By default the timeseries are updated. This behaviour can be changed by
using one of the other methods. See details.}
}
\value{
an updated \code{\link{regts}} object.
}
\description{
This function can be used to update, replace or extend a \code{(reg)ts}
object with another \code{(reg)ts} object.
The result is an updated \code{\link{regts}} object.
}
\details{
The two timeseries must have the same frequency, but may have a different
period range.
The common columns in the timeseries can be updated in four different ways:

\code{upd} the first timeseries are updated with the second timeseries for the
total period range of the second timeseries. Outside this period the values
in the first timeseries do not change.

\code{updna} if method \code{updna} is selected instead of \code{upd},
only NA values in the first timeseries will be updated

\code{updval} if method \code{updval} is selected instead of \code{upd},
the values in the first timeseries are only replaced with valid (i.e. non-NA)
values from the second timeseries.

\code{replace} like method \code{upd}, the values in the first
timeseries are replaced by the values in the second timeseries for the total
period range of these second timeseries.
Outside this period the values in the first timeseries will become
\code{NA}).

The non overlapping columns in both timeseries are added to the result.

The period range of the result is the union of the period ranges of the
first and second timeseries, except for the \code{updval} method.
For this method the result period range is the union of the first timeseries
and the timeseries obtained by applying function \code{na_trim} to the
second timeseries.
}
\examples{
x1 <- regts(matrix(data = rep(1:9), nc = 3), period = "2000/2002",
            names = c("a", "b", "c"))
x2 <- regts(matrix(data = rep(10:15), nc = 3), period = "2000/2001",
            names = c("a", "c", "d"))
update_ts(x1, x2, method = "upd")

}
\seealso{
\code{\link{regts}}
}
