% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_labels.R
\name{update_ts_labels}
\alias{update_ts_labels}
\title{Update one or more timeseries labels in a multivariate \code{\link{regts}} object}
\usage{
update_ts_labels(x, labels)
}
\arguments{
\item{x}{a multivariate \code{regts} object}

\item{labels}{a named character vector. The names are the names
of the timeseries (columns) whose label will be updated.
Specify \code{NULL} to remove all labels.}
}
\description{
Update one or more timeseries labels in a multivariate \code{\link{regts}} object
}
\examples{
ts <- regts(matrix(1:6, ncol = 2), start = "2016Q2", names = c("a", "b"),
             labels = c("Timeseries a", "???"))
ts <- update_ts_labels(ts, c(b = "Timeseries b"))
print(ts_labels(ts))

}
\seealso{
\code{\link{ts_labels}}
}
