% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ts.R
\name{write_ts_csv}
\alias{write_ts_csv}
\title{Write timeseries to a csv file}
\usage{
write_ts_csv(x, file, rowwise = TRUE, sep = ",", dec = ".",
  labels = c("after", "before", "no"), period_format = "regts",
  verbose = FALSE)
}
\arguments{
\item{x}{A \link{ts} or \link{regts} object.}

\item{file}{The filename of the output file.}

\item{rowwise}{A logical value: should the timeseries be written rowwise?}

\item{sep}{The separator between columns. Default is ",".}

\item{dec}{The decimal separator, by default ".". Cannot be the same as sep.}

\item{labels}{Should labels be written? If so, should it be
done before the names or after the names? By default,
labels are written after the names if present.}

\item{period_format}{The period format. By default the
\code{regts} format (e.g. \code{"2010Q2"}, see \code{\link{period}}) is used.
Alternatively, it is possible to specify a format employed by the base R
function \code{\link[base]{strptime}}, e.g. \code{"\%Y-\%m-\%d"}.}

\item{verbose}{A logical (default \code{FALSE}). If \code{TRUE}, the function
prints the filename, the number of timeseries written, the period range, and
the elapsed time.}
}
\description{
Write timeseries to a csv file, using \link[data.table:fwrite]{data.table::fwrite}.
}
\examples{
# create a timeseries object
ts1 <- regts(matrix(rnorm(50), ncol = 2), names = c("a", "b"),
             labels = c("Timeseries a", "Timeseries b"), start = "2017Q2")

# write timeseries to csv
write_ts_csv(ts1, file = "ts1.csv", labels = "after")

# write timeseries columnwise to csv, using a specified period_format
write_ts_csv(ts1, file = "ts1_2.csv", rowwise = FALSE,
             period_format = "\%Y-\%m-\%d")
\dontshow{
   unlink("ts1.csv")
   unlink("ts1_2.csv")
}
}
\seealso{
\link{read_ts_csv} and \link{write_ts_xlsx}.
}
