% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ts.R
\name{write_ts_csv}
\alias{write_ts_csv}
\title{Write timeseries to a csv file}
\usage{
write_ts_csv(x, file, rowwise = TRUE, sep = ",", dec = ".",
  labels = c("after", "before", "no"))
}
\arguments{
\item{x}{a \code{ts} or \code{regts} object}

\item{file}{a \code{regts} object}

\item{rowwise}{a logical value: should the timeseries we written rowwise?}

\item{sep}{The separator between columns. Default is ",".}

\item{dec}{The decimal separator, by default ".". Cannot be the same as sep.}

\item{labels}{should labels we written, and if so before the names or after
the names? By default, labels are written after the names if present.}
}
\description{
This function writes timeseries to a csv file.
The csv file is actually written by function \code{\link[data.table]{fwrite}}
of package \code{data.table}.
}
