% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ts.R
\name{write_ts_xlsx/write_ts_sheet}
\alias{write_ts_xlsx/write_ts_sheet}
\alias{write_ts_xlsx}
\alias{write_ts_sheet}
\title{Functions for writing timeseries to an xlsx file}
\usage{
write_ts_xlsx(x, file, sheet_name = "Sheet1", rowwise = TRUE,
  append = FALSE, labels = c("after", "before", "no"), comments,
  number_format, period_as_date = FALSE)

write_ts_sheet(x, sheet, rowwise = TRUE, labels = c("after", "before",
  "no"), comments, number_format, period_as_date = FALSE)
}
\arguments{
\item{x}{a \code{\link{ts}} or \code{\link{regts}} object}

\item{file}{the filename of the output file}

\item{sheet_name}{the sheet name}

\item{rowwise}{a logical value: should the timeseries be written rowwise?}

\item{append}{If \code{FALSE} (the default), then the original file,
if it exists, is replaced with the new file. All original data is lost.
If \code{TRUE}, then only data on the sheet with
the specified sheet name is erased and replaced with new data.  If the sheet
does not yet exist, then a new sheet is created and appended to the
original file.}

\item{labels}{should labels be written, and if so before or after
the names? By default, labels are written after the names if present}

\item{comments}{a character vector or data frame. The comments
are written to the beginning of the sheet, before the timeseries data is
written.}

\item{number_format}{a character value specifying the number format.
For example, \code{"#.00"} corresponds to two decimal spaces.
For details see the description of the function \code{\link[xlsx]{DataFormat}}
in the \code{\link[xlsx:xlsx-package]{xlsx}} package.}

\item{period_as_date}{A logical (default \code{FALSE}).
If \code{TRUE} the periods are written as date values to the Excel file.
By default the periods are written as characters using the standard
\code{regts} format (e.g. \code{"2010Q2"}, see \code{\link{period}}).}

\item{sheet}{a \code{\link[xlsx]{Sheet}} object (see the documentation
of package \code{xlsx})}
}
\description{
These functions can be used to write timeseries to a sheet of an
xlsx file. \code{write_ts_xlsx} creates or opens an Excel workbook
(depending on argument \code{append}) and writes the timeseries to
a sheet with a specified name.
\code{write_ts_sheet} writes timeseries to a \code{\link[xlsx]{Sheet}}
object.
}
\details{
The functions employ function \code{\link[xlsx]{addDataFrame}}
from the \code{\link[xlsx:xlsx-package]{xlsx}} package for writing the Excel file.

If you want to write multiple timeseries objects to different
sheets, you can use \code{write_ts_xlsx} with argument
\code{append = TRUE}, or \code{write_ts_sheet}. The latter approach
is more efficient.
}
\section{Functions}{
\itemize{
\item \code{write_ts_xlsx}: writes timeseries to an Excel
workbook

\item \code{write_ts_sheet}: writes a timeseries to a
\code{Sheet} object
}}

\examples{
# create two timeseries objects
ts1 <- regts(matrix(rnorm(50), ncol =  2), names = c("a", "b"),
             labels = c("Timeseries a", "Timeseries b"), start = "2017Q2")

# write timeseries ts1 to an Excel file
write_ts_xlsx(ts1, file = "ts1.xlsx", sheet_name = "ts1", labels = "after")

# write two sheets using write_ts_sheet
wb <- createWorkbook()
sheet <- createSheet(wb, "ts1")
write_ts_sheet(ts1, sheet, labels = "after")
sheet <- createSheet(wb, "ts1_times_100")
write_ts_sheet(ts1 * 100, sheet = sheet, labels = "after")
saveWorkbook(wb, "timeseries.xlsx")
#
# write a timeseries with comments
comments <- c("Timeseries ts1 is created on the Central Bureau of Policy Analysis",
              "using a random number generator")
write_ts_xlsx(ts1, file = "ts_comments.xlsx", sheet_name = "ts1",
              comments = comments)
\dontshow{
   unlink("ts1.xlsx")
   unlink("timeseries.xlsx")
   unlink("ts_comments.xlsx")
}
}
