% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ts.R
\name{write_ts_xlsx}
\alias{write_ts_xlsx}
\alias{write_ts_sheet}
\title{Write timeseries to an xlsx file}
\usage{
write_ts_xlsx(x, file, sheet_name = "Sheet1", rowwise = TRUE,
  append = FALSE, labels = c("after", "before", "no"), comments,
  number_format, period_as_date = FALSE, max_col_width = 60,
  verbose = FALSE)

write_ts_sheet(x, wb, sheet_name = "Sheet1", rowwise = TRUE,
  labels = c("after", "before", "no"), comments, number_format,
  period_as_date = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{A \code{\link{ts}} or \code{\link{regts}} object}

\item{file}{The filename of the output file}

\item{sheet_name}{The sheet name}

\item{rowwise}{A logical value: should the timeseries be written rowwise?}

\item{append}{If \code{FALSE} (the default), then the original file,
if it exists, is replaced with the new file. All original data is lost.
If \code{TRUE}, then only data on the sheet with
the specified sheet name is erased and replaced with new data.  If the sheet
does not yet exist, then a new sheet is created and appended to the
original file.}

\item{labels}{Should labels be written, and if so before or after
the names? By default, labels are written after the names if present}

\item{comments}{A character vector or data frame. The comments
are written to the beginning of the sheet, before the timeseries data is
written.}

\item{number_format}{A character value specifying the number format.
For example, \code{"#.00"} corresponds to two decimal spaces.
For details see the description of the function
\link[openxlsx:createStyle]{openxlsx::createStyle}.}

\item{period_as_date}{A logical (default \code{FALSE}).
If \code{TRUE} the periods are written as date values to the Excel file.
By default the periods are written as characters using the standard
\code{regts} format (e.g. \code{"2010Q2"}, see \code{\link{period}}).}

\item{max_col_width}{Integer (default 50). The column widths are
adjusted automatically, but are never larger than the specified value.}

\item{verbose}{A logical (default \code{FALSE}). If \code{TRUE}, the function
prints the file and sheet name, the number of timeseries written, the period
range, and the elapsed time.}

\item{wb}{A \code{Workbook} object created with function
\link[openxlsx:createWorkbook]{openxlsx::createWorkbook} or \link[openxlsx:loadWorkbook]{openxlsx::loadWorkbook}.}
}
\description{
\code{write_ts_xlsx} creates or opens an Excel workbook
(depending on argument \code{append}) and writes the timeseries to
a sheet with a specified name.
\code{write_ts_sheet} writes timeseries to a sheet of a workbook
object returned by function \link[openxlsx:createWorkbook]{openxlsx::createWorkbook}
or \link[openxlsx:loadWorkbook]{openxlsx::loadWorkbook}.
}
\details{
The functions employ package \link[openxlsx:openxlsx]{openxlsx::openxlsx} for writing the Excel file.

If you want to write multiple timeseries objects to different
sheets, you can use \code{write_ts_xlsx} with argument
\code{append = TRUE}. Alternatively,
you can create a workbook object with
function \link[openxlsx:createWorkbook]{openxlsx::createWorkbook} of package
\link[openxlsx:openxlsx]{openxlsx::openxlsx} and then add a sheet with \code{write_ts_sheet}.
The latter approach is more efficient.
When the workbook is written to a file with function
\link[openxlsx:saveWorkbook]{openxlsx::saveWorkbook}, it is often useful to
set the minimum and maximum scolumn width option for package \link[openxlsx:openxlsx]{openxlsx::openxlsx},
as shown in the example below.
}
\section{Functions}{
\itemize{
\item \code{write_ts_xlsx()}: writes timeseries to an Excel workbook file.

\item \code{write_ts_sheet()}: writes a timeseries to a workbook object
returned by \link[openxlsx:createWorkbook]{openxlsx::createWorkbook} or \link[openxlsx:loadWorkbook]{openxlsx::loadWorkbook}.

}}
\section{Warning}{

When using \code{write_ts_xlsx} with \code{append = TRUE},
formulas on existing sheets are \strong{not} reevaluated. The same applies when
\code{write_ts_sheet} is used to add a sheet to an existing workbook.
Open the file in Excel and press \code{F9} to recalculate all formulas manually.
}

\examples{
# create a timeseries object
ts1 <- regts(matrix(rnorm(50), ncol =  2), names = c("a", "b"),
             labels = c("Timeseries a", "Timeseries b"), start = "2017Q2")

# write timeseries ts1 to an Excel file
write_ts_xlsx(ts1, file = "ts1.xlsx", sheet_name = "ts1", labels = "after",
              verbose = TRUE)

# write two sheets using write_ts_sheet
library(openxlsx)
wb <- createWorkbook()
write_ts_sheet(ts1, wb, "ts1", labels = "after")
write_ts_sheet(ts1 * 100, wb, "ts1_times_100", labels = "after",
               verbose = TRUE)

# Set the minimum and maximum column width, to prevent very narrow or wide
# columns. saveWorkbook will adjust the widths of the columns written by
# write_ts_sheet.
options(openxlsx.minWidth = 8.43, openxlsx.maxWidth = 60)

# Save the workbook with openxlsx::saveWorkbook.
ok <- saveWorkbook(wb, "timeseries.xlsx", overwrite = TRUE,
                   returnValue = TRUE)
if (!ok) {
  stop("Failed to save workbook to file 'timeseries.xlsx'. Check warnings.")
}

# write a timeseries with comments
comments <- c("Timeseries ts1 is created at the CPB",
              "using a random number generator")
write_ts_xlsx(ts1, file = "ts_comments.xlsx", sheet_name = "ts1",
              comments = comments)
\dontshow{
   unlink("ts1.xlsx")
   unlink("timeseries.xlsx")
   unlink("ts_comments.xlsx")
}
}
\seealso{
\link{read_ts_xlsx} and \link{write_ts_csv}.
}
