% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ts.R
\name{write_ts_xlsx}
\alias{write_ts_xlsx}
\title{Write timeseries to an xlsx file}
\usage{
write_ts_xlsx(x, file, sheet_name = "Sheet1", rowwise = TRUE,
  labels = c("after", "before", "no"))
}
\arguments{
\item{x}{a \code{ts} or \code{regts} object}

\item{file}{a \code{regts} object}

\item{sheet_name}{the sheet name}

\item{labels}{should labels we written, and if so before the names or after
the names? By default, labels are written after the names if present.}
}
\description{
This function writes timeseries to an xlsx  file.
The xlsx file is actually written by function \code{\link[xlsx]{write.xlsx}}
of package \code{data.table}.
}
