% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ts.R
\name{write_ts_xlsx}
\alias{write_ts_xlsx}
\alias{write_ts_sheet}
\title{Functions for writing timeseries to an xlsx file}
\usage{
write_ts_xlsx(x, file, sheet_name = "Sheet1", rowwise = TRUE,
  append = FALSE, labels = c("after", "before", "no"))

write_ts_sheet(x, wb, sheet_name = "Sheet1", rowwise = TRUE,
  labels = c("after", "before", "no"))
}
\arguments{
\item{x}{a \code{ts} or \code{regts} object}

\item{file}{the filename of the output file}

\item{rowwise}{a logical value: should the timeseries we written rowwise?}

\item{append}{If \code{FALSE} (the default), then the original file,
if it exists, is replaced with the new file. All original data is lost.
If \code{TRUE}, then only data on the sheet with
the specified sheet name is erased and replaced with new data.  If the sheet
does not yet exist, then a new sheet is created and appended to the
original file.}

\item{labels}{should labels we written, and if so before the names or after
the names? By default, labels are written after the names if present}

\item{wb}{a workbook object as returned by \code{createWorkbook}
or \code{loadWorkbook}.}
}
\description{
These functions can be used to write timeseries to an xlsx file.
The function employ package  \code{\link[xlsx]{write.xlsx}} for
writing the Excel file.
}
\section{Functions}{
\itemize{
\item \code{write_ts_sheet}: Writes a timeseries to a sheet of an Excel workbook
}}

